% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{app_ui}
\alias{app_ui}
\title{App ui}
\usage{
app_ui(request)
}
\arguments{
\item{request}{Internal parameter for '`shiny`'.
DO NOT REMOVE.}
}
\description{
The application User-Interface (UI).
}
\details{
The application User-Interface
}
