% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{c_argmatch}
\alias{c_argmatch}
\title{Match arguments}
\usage{
c_argmatch(x, y)
}
\arguments{
\item{x}{the argument to match fuzzily}

\item{y}{a character vector containing the options}
}
\value{
The index of the matching value (NA if no valid match)
}
\description{
find the best partial match for x in y, ignoring case
}
