% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityBootstrap.R
\name{c_BootstrapValue}
\alias{c_BootstrapValue}
\title{Compute Residuals or Persistences on bootstrapped data without scaling}
\usage{
c_BootstrapValue(clearned, Ybs, summary = abs, diag = 0, ...)
}
\arguments{
\item{clearned}{either a ClarityScan or Clarity object. If a Clarity object, then residuals will be bootstrapped. If a ClarityScan object then Persistence will be bootstrapped, unless a specific k is given in which case Residuals are again calculated.}

\item{Ybs}{The bootstrapped matrix for which residuals will be calculated}

\item{summary}{(default=abs) Summarisation of the bootstrapped value when extracting Residuals. Default: abs, which allows residuals to be easily compared. Use summary=I to return the raw data. See \code{\link{Clarity_Extract}}.}

\item{diag}{(default=0) diagonal value to be applied when extracting Residuals. See \code{\link{Clarity_Extract}}.}

\item{...}{Additional parameters to \code{\link{Clarity_Persistence}} (not required for normal usage)}
}
\value{
A list of residuals if clearned is of class Clarity, or a list of persistences if clearned is of class ClarityScan.
}
\description{
Given a resampled matrix Ybs which was originally used to make a Clarity or ClarityScan object clearned, we:
1. Use the provided clearned object to predict the bootstrapped data
2. Compute the residual structure of this prediction

We predict the target with \code{\link{Clarity_Predict}}.
}
\seealso{
\code{\link{Clarity_Bootstrap}} which is the recommended way to access this code.
}
