% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clarity.R
\name{c_calXfromX}
\alias{c_calXfromX}
\title{Obtain normalised calX from X}
\usage{
c_calXfromX(A, X)
}
\arguments{
\item{A}{An N by K non-negative matrix of mixture proportions}

\item{X}{A K by K non-negative matrix of latent weights}
}
\value{
A list containing the following:
\itemize{
\item C, the matrix C above
\item tCinv the matrix t(C)^{-1}
\item calX non-negative matrix of dimension K by K
}
}
\description{
This function moves between a fit represented as
     Y = || A X A^T ||_2
to one represented as
     Y = || A calX Lambda^T ||_2
where calX now is on the same scale as Y, and Lambda is A with columns summing to 1.  calX is the "naturally scaled" version of X

We have X = calX t(C)
So calX = X (t(C))^{-1}

Where t(C) is defined by t(Lambda) = t(C) t(A), the matrix which makes Lambda be A with columns summing to 1
}
\keyword{mixture}
