% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityTransforms.R
\name{c_Center}
\alias{c_Center}
\title{(Double) Center a matrix}
\usage{
c_Center(Y, col = TRUE, row = TRUE, method = "mean", na.rm = TRUE)
}
\arguments{
\item{Y}{the matrix to be centered}

\item{col}{(default: TRUE) Whether to center the columns to ensure they have mean zero}

\item{row}{(default: TRUE) Whether to center the rows to ensure they have mean zero}

\item{method}{Either "mean" (Default) or "matrix"; Whether to center with the theoretically nice centering matrix \code{\link{c_getC}} or just subtracting row and column means appropriately, which is quicker, and supports na values.}

\item{na.rm}{Whether to remove NA values if method=="mean"}
}
\value{
A matrix with same dimensions as Y, centered.
}
\description{
Centers a matrix by rows and/or columns, nominally computing 'Cn %*% Y %*% Cm' for an n by m matrix Y
}
