% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityBootstrap.R
\name{c_CompareNormal}
\alias{c_CompareNormal}
\title{Compare Observed to Bootstrap values (Normal approximation one tailed test)}
\usage{
c_CompareNormal(x, o)
}
\arguments{
\item{x}{the distribution of theoretical or simulated values}

\item{o}{the observed value}
}
\value{
a p value
}
\description{
Compare an observation (o) with a set of values (x) and report the p value for a one tailed test that o is large compared to x. This is usually the most appropriate test in Clarity.

If the number of observations is high, consider \code{\link{c_CompareEmpirical}} instead.

For a two-tailed test, consider \code{\link{c_CompareNormalTwoTailed}} instead.
}
\seealso{
\code{\link{c_pval}} in which this function can be used.
}
