% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityBootstrap.R
\name{c_CompareNormalTwoTailed}
\alias{c_CompareNormalTwoTailed}
\title{Compare Observed to Bootstrap values (Normal approximation two tailed test)}
\usage{
c_CompareNormalTwoTailed(x, o)
}
\arguments{
\item{x}{the distribution of theoretical or simulated values}

\item{o}{the observed value}
}
\value{
a p value
}
\description{
Compare an observation (o) with a set of values (x) and report the p value for a one tailed test that o is far from the mean of x.

If the number of observations is high, consider \code{\link{c_CompareEmpiricalTwoTailed}} instead.

For a one-tailed test, consider \code{\link{c_CompareNormal}} instead.
}
\seealso{
\code{\link{c_pval}} in which this function can be used.
}
