% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityCompare.R
\name{c_crossvalidate}
\alias{c_crossvalidate}
\title{Cross validatation method to estimate the number of PCs associated with structure in a Ylist}
\usage{
c_crossvalidate(datalist, kmax = NULL, ...)
}
\arguments{
\item{datalist}{A list of pairs of N by N matrices, each independently sampled from the same data source as described in \code{\link{Clarity_Predict}}.}

\item{kmax}{(default=NULL, meaning use the dimension of the data)}

\item{...}{additional parameters to \code{link{Clarity_Scan}}.}
}
\value{
An list containing:
\itemize{
\item min.score : the point estimate of kmax, obtained from the minimum average prediction error.
\item scores : the average prediction error at each complexity \code{K}.
\item softk : For each complexity \code{K}, the probability that the estimated \code{K^*} is at least as large as \code{K}.
}
}
\description{
From a list of pairs of resampled matrices with the same distribution, use the first to predict the second. The prediction error is then used to form an estimate of the number of dimensions that are increasing predictive power.

This is summarised in 2 ways. Firstly, a point estimate of kmax: the minimum of the average prediction error. Secondly, a soft estimate of Kmax: the proportion of samples that have a minimum prediction error larger than a given K.
}
\seealso{
\code{\link{Clarity_Compare}}, which takes an argument distfn which can be given this or another function that generates a similarity from feature data.
}
