% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityPlots.R
\name{c_extract}
\alias{c_extract}
\title{Extract objects from a ClarityScan object}
\usage{
c_extract(cscan, what = "Yresid")
}
\arguments{
\item{cscan}{A ClarityScan object}

\item{what}{(default="Yresid") the object to extract}
}
\value{
A list containing what for each k in the ClarityScan
}
\description{
Extract objects such as the Yresid residuals from a ClarityScan object. This is useful for constructing Persistence Diagrams or similar.
}
\seealso{
\code{\link{c_Persistence}} to which you can pass
}
