% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityCompare.R
\name{c_fixdiagonals}
\alias{c_fixdiagonals}
\title{Diagonal standardization for similarity/dissimilarity matrices}
\usage{
c_fixdiagonals(x, method = "detect")
}
\arguments{
\item{x}{An N by N matrix}

\item{method}{(default="detect") standardization method; the default detects whether a similarity or dissimilarity is provided and uses the max/min respectively. Can be manually set to the function min,max or other operation on a row,column tuple.}
}
\value{
An N by N matrix of the distances
}
\description{
Complete the diagonal with a sensible approximation that will work for SVD and significance testing.
}
