% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clarity.R
\name{c_ForwardBackward}
\alias{c_ForwardBackward}
\title{Update a Clarity estimate by scanning over K}
\usage{
c_ForwardBackward(clist, niter = 10, thresh = -1, tmax = 10000,
  verbose = TRUE, matrixdistmin = 1e-08, objectivedistmin = 0.001,
  ...)
}
\arguments{
\item{clist}{A list of Clarity objects as returned by \code{\link{Clarity_Scan}}}

\item{niter}{(default 10) maximum number of forward-backward iterations}

\item{thresh}{(default -1) required improvement in objective function across all K before we give up. If negative, then we always try niter iterations. Setting a small, positive value can speen computation in large problems.}

\item{tmax}{tmax as passed to \code{\link{Clarity_fixedK}}}

\item{verbose}{Whether to output information about progress. Note that Clarity_fixedK is always run with verbose=FALSE}

\item{matrixdistmin}{(default NULL meaning derive from clist) Passed to \code{\link{Clarity_fixedK}}.}

\item{objectivedistmin}{(default NULL meaning derive from clist) Passed to \code{\link{Clarity_fixedK}}.}

\item{...}{futher parameters to be passed to \code{\link{Clarity_fixedK}}}
}
\value{
A object of class "ClarityScan" as returned by \code{\link{Clarity_Scan}}
}
\description{
Takes a list of Clarity results at different K and update them all, by iteratively trying to improve lower K estimates from higher K, and vice-versa
}
\seealso{
\code{\link{Clarity_Scan}} is the recommended interface for performing  inference on a representation of a single similarity matrix with Clarity.
}
\keyword{mixture}
