% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clarity.R
\name{c_ForwardStep}
\alias{c_ForwardStep}
\title{Update a Clarity estimate by scanning over increasing K}
\usage{
c_ForwardStep(clist, verbose = TRUE, tmax = 10000, matrixdistmin = NULL,
  objectivedistmin = NULL, ...)
}
\arguments{
\item{clist}{A list of Clarity objects as returned by \code{\link{Clarity_Scan}}}

\item{verbose}{Whether to output information about progress. Note that Clarity_fixedK is always run with verbose=FALSE}

\item{tmax}{tmax as passed to \code{\link{Clarity_fixedK}}}

\item{matrixdistmin}{(default NULL meaning derive from clist) Passed to \code{\link{Clarity_fixedK}}.}

\item{objectivedistmin}{(default NULL meaning derive from clist) Passed to \code{\link{Clarity_fixedK}}.}

\item{...}{futher parameters to be passed to \code{\link{Clarity_fixedK}}}
}
\value{
A object of class "ClarityScan" as returned by \code{\link{Clarity_Scan}}
}
\description{
Takes a list of Clarity results at different K and update them all, using estimates at lower K to improve those at higher K.
}
\seealso{
\code{\link{Clarity_Scan}} is the recommended interface for performing  inference on a representation of a single similarity matrix with Clarity.
}
\keyword{mixture}
