% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClaritySVD.R
\name{c_GetAFromSimplex}
\alias{c_GetAFromSimplex}
\title{Compute locations of data inside a simplex}
\usage{
c_GetAFromSimplex(Z, ZS)
}
\arguments{
\item{Z}{locations to be fitted, N rows with k columns}

\item{ZS}{enclosing simplex: k+1 rows with k columns}
}
\value{
A mixture matrix containing N rows summing to 1, with k+1 columns
}
\description{
Given data Z and an enclosing simplex ZS, compute the locations of Z. Uses a linear model (lm) to do the computation
}
