% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityTransforms.R
\name{c_getC}
\alias{c_getC}
\title{Get a centering matrix of specified dimension}
\usage{
c_getC(n)
}
\arguments{
\item{n}{either the dimension of Y, or Y itself.}
}
\value{
A matrix with same dimensions as Y, centered.
}
\description{
Gets a centering matrix Cn so that Cn %*% Y is centered.
}
