% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityPlots.R
\name{c_getplotval}
\alias{c_getplotval}
\title{Get a plot value in a sensible way}
\usage{
c_getplotval(x, i, clist, default)
}
\arguments{
\item{x}{the provided value}

\item{i}{the index}

\item{clist}{the list of clarityscan objects}

\item{default}{what to return if x is null}
}
\value{
the desired plot symbol
}
\description{
Generate a default colour/line type/pch etc for a ClarityScan object. Users are suggested to supply their own choices rather than call c_getplotval directly.
}
