% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityTransforms.R
\name{c_Identity}
\alias{c_Identity}
\title{Identity function taking two arguments}
\usage{
c_Identity(Ystart, Ytarget = NULL)
}
\arguments{
\item{Ystart}{the matrix to transform}

\item{Ytarget}{The target matrix}
}
\value{
Ystart, unmodified
}
\description{
Returns the Y matrix untransformed. Provided for input as the transform variable in \code{\link{c_Bootstrap}}.
}
\seealso{
\code{\link{c_Bootstrap}}, which uses this code.
}
