% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityMultiplicative.R
\name{c_IncreaseK}
\alias{c_IncreaseK}
\title{Make a larger Clarity object by randomly peturbing a smaller one}
\usage{
c_IncreaseK(Y, AatKm1, XatKm1, alpha = 0.9, dirichletbeta = 1)
}
\arguments{
\item{Y}{Non-negative Similiarty matrix being fitted (N by N)}

\item{AatKm1}{Non-negative matrix of mixtures (N by K)}

\item{XatKm1}{Non-negative matrix of latent structures (K by K)}

\item{alpha}{(default 0.9) Mixing weight between original matrix and new matrix}

\item{dirichletbeta}{(default 1) Dirichlet beta for the new random A being mixed with the original}
}
\value{
A list containing the following:
\itemize{
\item X A K by K non-negative matrix of latent weights
\item A An N by K non-negative matrix of mixture proportions
\item dist The objective function evaluated at the new A and X
}
}
\description{
Make a Clarity object of size K+1 from one of size K by mixing in a random matrix with the old one
}
\keyword{mixture}
