% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clarity.R
\name{c_initKmeans}
\alias{c_initKmeans}
\title{Initialise A and X in Clarity using Kmeans}
\usage{
c_initKmeans(Y, K, a = 0.9, sigma = 1e-04, sigmaY = 1e-05,
  nstart = 1000, iter.max = 50)
}
\arguments{
\item{Y}{An N by N non-negative matrix of similarities}

\item{K}{number of clusters, passed to \code{\link{kmeans}}}

\item{a}{(default 0.9) Fraction of mixture to be retained in the chosen cluster}

\item{sigma}{(default 0.0001) Multiplicative noise for A}

\item{sigmaY}{(default 1e-5) Multiplicative noise for Y, to prevent co-linearity}

\item{nstart}{(default 1000) Number of restarts, passed to \code{\link{kmeans}}}

\item{iter.max}{(default 50) Number of iterations, passed to \code{\link{kmeans}}}
}
\value{
A list containing the following:
\itemize{
\item Y An N by N non-negative matrix of similarities
\item A An N by K non-negative matrix of mixture proportions
\item X A K by K non-negative matrix of latent weights
}
}
\description{
This function takes a non-negative similarity matrix Y of dimension N by N, 
and a desired number of clusters K,
and uses K-means to estimate them.

It then "noises" these clusters by retaining a membership fraction a of the chosen cluster and distributing membership (1-a) amongst the other K-1 clusters, with additional multiplicative noise N(1,sigma).
}
\seealso{
\code{\link{Clarity_Scan}} is the recommended interface for performing  inference on a representation of a single similarity matrix with Clarity.
}
\keyword{mixture}
