% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClaritySimpleSVD.R
\name{c_IterativeSVD}
\alias{c_IterativeSVD}
\title{Iteratively apply an SVD to remove the effect of diagonals}
\usage{
c_IterativeSVD(Y, K, Tmax = 25, tol = 1e-05)
}
\arguments{
\item{Y}{The similarity data}

\item{K}{The SVD complexity to retain}

\item{Tmax}{(Default: 25) Maximum number of iterations to process. Typically converges quite quickly.}

\item{tol}{(Default: 1e-5) Tolerance of the difference in the sum of the residuals on the diagonal.}
}
\value{
A list containing:
\itemize{
\item Yp.svd: the SVD object of (Y-diag(D));
\item D: the estimated D;
\item K: the provided K;
\item Dhistory: a matrix of dimension (t times n) of the diagonals at each iteration i;
\item reshistory: a vector of the squared residuals at each iteration i;
\item Tmax: The final Tmax, after any tolerance is applied.
}
}
\description{
The self-similarity or difference is often problematic, or follows a different model to the pairwise values. This function applies SVD and computes residuals on the diagonal, removes these residuals, and iterates. The resulting object will minimize Y - X X^T
}
