% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityPlots.R
\name{c_listDivergence}
\alias{c_listDivergence}
\title{Compute the standardized divergence between a prediction and the model from which it was built}
\usage{
c_listDivergence(y2pred, y1scan)
}
\arguments{
\item{y2pred}{A ClarityScan object produced by Clarity_Predict, predicted from y1scan}

\item{y1scan}{A ClarityScan object produced by Clarity_Scan}
}
\value{
A list of matrices, each containing the Standardized Divergence Matrix
}
\description{
Compute D(Y_1,Y_2,K) - (Y_1-Y_2) in the notation of Lawson et al 2019 CLARITY paper
}
\seealso{
\code{\link{c_listSymmetricDivergence}} to compute the difference betwe
}
