% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityBootstrap.R
\name{c_listSymmetricDivergence}
\alias{c_listSymmetricDivergence}
\title{Compute the standardized symmetric divergence between a prediction and the model from which it was built}
\usage{
c_listSymmetricDivergence(y1scan, y2scan, y12pred, y21pred)
}
\arguments{
\item{y1scan}{A ClarityScan object produced by Clarity_Scan}

\item{y2scan}{A ClarityScan object produced by Clarity_Scan}

\item{y12pred}{A ClarityScan object produced by Clarity_Predict, predicted from y2scan}

\item{y21pred}{A ClarityScan object produced by Clarity_Predict, predicted from y1scan}
}
\value{
A list of matrices, each containing the Standardized Divergence Matrix
}
\description{
Compute
[D(Y_1,Y_2,K) - (Y_1-Y_2)]-[D(Y_2,Y_1,K) - (Y_2-Y_1)]
= D(Y_1,Y_2,K) - D(Y_2,Y_1,K)
in the notation of Lawson et al 2019 CLARITY paper
}
\seealso{
\code{\link{c_listDivergence}} which is used to compute the divergences.
}
