% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityTransforms.R
\name{c_MeanScale}
\alias{c_MeanScale}
\title{Perform a Mean Scale transformation of one matrix into another}
\usage{
c_MeanScale(Ystart, Ytarget)
}
\arguments{
\item{Ystart}{the matrix to transform}

\item{Ytarget}{The target matrix}
}
\value{
A matrix which is the best fitting transformation of Ystart to Ytarget
}
\description{
Scale the mean of the matrix provided in the first argument, into the target provided in the second.
}
\seealso{
\code{\link{c_Bootstrap}}, which uses this code.
}
