% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{c_Merge}
\alias{c_Merge}
\title{Merge a rectangular matrix into a lower dimensional matrix using mixtures}
\usage{
c_Merge(x, A, rotate = FALSE, type = "mean")
}
\arguments{
\item{x}{a matrix of dimension N by P}

\item{A}{a mixture matrix  of dimension N by K}

\item{rotate}{(default=FALSE) whether to return a rotated version of the result (by omitting the final transpose)}

\item{type}{(default="mean") whether to reaturn the "mean" or the "sum" of pairwise elements across the (soft) clustering described by A}
}
\value{
A matrix of dimension N by K
}
\description{
Takes a matrix x (N by P) and a mixture matrix A (N by K) and computes an N by K matrix:
\code{t(t(x) \%*\% A)}
Also has the facility to avoid the final transpose, and to replace A by normalised version that switches the computation from a weighted sum across "clusters" to a weighted mean by using
\code{An=t(t(A)/colSums(A))}
instead of A.
}
\seealso{
\code{\link{plot.ClarityScan}} usez this to make a summary of Clarity persistences.
}
