% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityMultiplicative.R
\name{c_multiplicative_fixedK}
\alias{c_multiplicative_fixedK}
\title{Fit a Similarity matrix using Clarity}
\usage{
c_multiplicative_fixedK(Y, K = NULL, tmax = 10000,
  matrixdistmin = 1e-08, objectivedistmin = 0.001, printskip = 100,
  verbose = TRUE, updateA = TRUE, updateX = TRUE, A = NULL,
  X = NULL, a = 0.9, sigma = 1e-04, sigmaY = 1e-05,
  ensurepositive = TRUE, fixed = NULL, ...)
}
\arguments{
\item{Y}{An N by N non-negative matrix of similarities}

\item{K}{an integer giving the dimension of the fit to be used}

\item{tmax}{(default 1e4) how many iterations to run for. maximimum}

\item{matrixdistmin}{(default 1e-8) stopping criterion for the summed square difference of A and X}

\item{objectivedistmin}{(default 1e-3) stopping criterion for the increase in objective function}

\item{printskip}{(default 100) how many iterations between progress updates}

\item{verbose}{(default TRUE) whether to output any information regarding progress}

\item{updateA}{(default TRUE) whether to update A; shouldn't use this option}

\item{updateX}{(default TRUE) whether to update X; shouldn't use this option}

\item{A}{(default NULL) initial A. Must pass A and X together to specify}

\item{X}{(default NULL) initial X. Must pass A and X together to specify}

\item{a}{(default 0.9) a parameter for \code{\link{c_initKmeans}}}

\item{sigma}{(defailt 0.0001) a parameter for \code{\link{c_initKmeans}}}

\item{sigmaY}{(defailt 1e-5) a parameter for \code{\link{c_initKmeans}}}

\item{ensurepositive}{(default TRUE) Whether we restrict to non-negative Y. Note that the algorithm can work with some negative Y as long as it is postitive semi-definite, but this can still lead to convergence problems.}

\item{fixed}{(efault NULL) which columns to keep structurally fixed. Not yet used.}
}
\value{
A list of class "Clarity" containing the following:
\itemize{
\item Y, the provided non-negative N by N similarity matrix
\item A, the inferred non-negative N by K mixture matrix
\item X, the inferred non-negative K by K latent matrix
\item calX, the inferred non-negative K by K 'naturally scaled' latent matrix
\item prediction, A X A^T
\item C, the matrix C above relating X to calX
\item tCinv, the inverse of C^T, used in relating X to calX
\item Afirst, the initial estimate of A
\item Xfirst, the initial estimate of X
\item calXfirst, the initial estimate of calX
\item numiters, the number of actually used iterations
\item matrixdist, the distance between matrices, a vector of length numiters
\item objective, the objective function at the final (A,X;Y)
\item objectivedist, the objective function, a vector of length numiters
\item matrixdistmin, the matrix change threshold provided
\item objectivedistmin, the objective change threshold provided
}
}
\description{
This function fits the best solution of A and X to
     || Y - A X A^T ||_2
for a FIXED dimension of the latent space K. Here,
A is n N by K non-negative matrix of mixture proportions
and
X is a K by K non-negative matrix of latent weights
}
\seealso{
\code{\link{Clarity_Scan}} is the recommended interface for performing  inference on a representation of a single similarity matrix with Clarity.
}
\keyword{mixture}
