% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClaritySimpleSVD.R
\name{c_nodiagSVD_Scan}
\alias{c_nodiagSVD_Scan}
\title{Compute the Clarity model by SVD embedding data for a range of model sizes}
\usage{
c_nodiagSVD_Scan(
  Y,
  kmax = NULL,
  Ysvd = NULL,
  verbose = TRUE,
  Xtype = "X",
  Tmax = 25,
  tol = 1e-05
)
}
\arguments{
\item{Y}{matrix to be fit}

\item{kmax}{(default: NULL, meaning use the data dimension) the maximum number of mixture components to use}

\item{Ysvd}{(default: NULL, meaning calculate it) the SVD of Y. If you provide a different object containing locations in the list Ysvd$u, they will be used instead. IMPORTANT: Ysvd$d must be rotated to ensure that Ysvd$u = Ysvd$v}

\item{verbose}{(default TRUE) whether to output progress to the terminal.}

\item{Xtype}{(default "X") whether to construct X by optimal solution ("X") or by SVD Sigma ("Sigma"). They are equivelent when Y is symmetric.}

\item{Tmax}{(default: 25) As passed to \code{\link{c_IterativeSVD}}.}

\item{tol}{(default: 1e-5) As passed to \code{\link{c_IterativeSVD}}.}
}
\value{
A ClarityScan object as described in \code{\link{Clarity_Scan}}, with the additional elements:
\itemize{
\item Ysvd The list of SVDs that was used in the embedding.
}
}
\description{
Given data Y, embed it into a space of dimension 1..kmax and fit it as an SVD
}
\seealso{
\code{\link{Clarity_Scan}} for the recommended interface.
}
