% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityPlots.R
\name{c_Persistence}
\alias{c_Persistence}
\title{Perform an operation on a list of matrices in order to compute a Persistence Structure}
\usage{
c_Persistence(mlist, f = function(x) rowSums(x^2))
}
\arguments{
\item{mlist}{A list of matrices}

\item{f}{Either a character string matching "Frobenius" or "RowSumsSquared", or a function to compute for each matrix, which should return a vector when operating on a matrix. Frobenius is the Frobenius norm of the matrix (sum(x^2)) and RowSumsSquared (default) is rowSums(x^2)}
}
\value{
A matrix of f returned for each matrix
}
\description{
Compute
[D(Y_1,Y_2,K) - (Y_1-Y_2)]-[D(Y_2,Y_1,K) - (Y_2-Y_1)]
= D(Y_1,Y_2,K) - D(Y_2,Y_1,K)
in the notation of Lawson et al 2019 CLARITY paper
}
