% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityCompare.R
\name{c_px2_lt_x1}
\alias{c_px2_lt_x1}
\title{Emprical estimator for comparing distributions}
\usage{
c_px2_lt_x1(x1, x2, reg = 1)
}
\arguments{
\item{x1}{A set of samples from the reference distribution}

\item{x2}{A set of samples from the query distribution}

\item{reg}{(Default=1) regularisation; add a null sample with this value to ensure that zero probability is not assigned to very small values.}
}
\value{
the probability that a sample from the query is larger than the reference.
}
\description{
the probability that a sample of a distribution (x2) is less than or equal to the sample of another distribution (x1), using their emprical distribution functions. A regularisation is included, to prevent 0s.

Probability that x2>=x1 is
sum_{i=1}^N pr(x1=x) pr(x2<=x)
= (1/(reg+length(x))) e2(c(reg,x1))
}
\examples{
c_px2_lt_x1(0,1) # 0.5 due to regularization
c_px2_lt_x1(0,1,0) # 0
c_px2_lt_x1(0:2,1:3,0) # 1/3
}
\seealso{
\code{\link{Clarity_Compare}}, the recommended way to access this code.
}
