% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClaritySVD.R
\name{c_regSimplex}
\alias{c_regSimplex}
\title{Fit a simplex around a ball}
\usage{
c_regSimplex(k)
}
\arguments{
\item{k}{Number of dimensions of the ball}
}
\value{
A matrix with each vertex as a column, of dimenison k+1
}
\description{
construct regular k simplex in R^k with barycenter the origin
related to square of the minimum distance of barycenter to origin
}
\seealso{
\code{\link{c_scaleSimplex}} to specify the radius
}
