% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClaritySVD.R
\name{c_scaleSimplex}
\alias{c_scaleSimplex}
\title{Fit a simplex around a ball of specified radius}
\usage{
c_scaleSimplex(k, r)
}
\arguments{
\item{k}{Number of dimensions of the ball}

\item{r}{Radius of the ball}
}
\value{
A matrix with each vertex as a column, of dimenison k+1
}
\description{
construct regular k simplex in R^k with barycenter the origin and radius r
related to square of the minimum distance of barycenter to origin
}
\seealso{
\code{\link{c_SimplexSurrounding}} to compute the ball radius from data
}
