% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityMultiplicative.R
\name{c_ScanOnly}
\alias{c_ScanOnly}
\title{Scan over K to learn a list of Clarity objects}
\usage{
c_ScanOnly(Y, kmax = 20, tmax = 10000, matrixdistmin = 1e-08,
  objectivedistmin = 0.001, printskip = 1000, verbose = TRUE, ...)
}
\arguments{
\item{Y}{An N by N non-negative matrix of similarities}

\item{kmax}{an integer giving the maximum dimension of the fit to be used (2..kmax are used)}

\item{tmax}{(default 1e4) how many iterations to run for. maximimum}

\item{matrixdistmin}{(default 1e-8) stopping criterion for the summed square difference of A and X}

\item{objectivedistmin}{(default 1e-3) stopping criterion for the increase in objective function}

\item{printskip}{(default 100) how many iterations between progress updates}

\item{verbose}{(default TRUE) whether to output any information regarding progress}

\item{...}{futher parameters to be passed to \code{\link{Clarity_fixedK}}}
}
\value{
A object of class "ClarityScan" as returned by \code{\link{Clarity_Scan}}
}
\description{
Runs Clarity_fixedK for all K up to some threshold. Returns a list of results which is improved upon by many other functions.
}
\seealso{
\code{\link{Clarity_Scan}} is the recommended interface for performing  inference on a representation of a single similarity matrix with Clarity.
}
\keyword{mixture}
