% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityCompare.R
\name{c_Scoresplit}
\alias{c_Scoresplit}
\title{Comparison for lists of pairs of matrices}
\usage{
c_Scoresplit(scorelist, comparison = c_px2_lt_x1)
}
\arguments{
\item{scorelist}{A list in which each element is a length 2 list, consisting of [X,Y] where X and Y are each matrices of equal dimension.}

\item{comparison}{(default=c_px2_gt_x1) comparison function to evaluate.  \code{function(d1,d2)=wilcox.test(d1,d2)$p.value} and similar distributional tests are more stringent but are not appropriate for resampled data.}
}
\value{
the probability that a sample from the query is larger than the reference.
}
\description{
Perform an element-wise comparison on the set of elements at a given location in a matrix, computing:
\code{comparison({X[i,j]}_{1..B},{Y[i,j]}_{1..B})}
i.e comparing the set of [i,j]-th elements of Y to the same set for X.

The intended use case is when X and Y are each generated by some scoring function for resamples of a reference dataset (X) and a query dataset (Y). The default uses \code{\link{c_px2_lt_x1}} to calculate the probability that Y[i,j] is less than X[i,j].
}
