% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClaritySimpleSVD.R
\name{c_simpleSVD_fixedK}
\alias{c_simpleSVD_fixedK}
\title{Compute the Clarity model by SVD embedding data for a fixed size of model k}
\usage{
c_simpleSVD_fixedK(Ysvd, Y, k, verbose = TRUE, Xtype = "X")
}
\arguments{
\item{Ysvd}{SVD of Y}

\item{Y}{matrix to be fit}

\item{k}{the number of SVD dimensions to retain}

\item{verbose}{(default TRUE) whether to output progress to the terminal.}

\item{Xtype}{(default "X") whether to construct X by optimal solution ("X") or by SVD Sigma ("Sigma"). They are equivelent when Y is symmetric.}
}
\value{
A Clarity object as described in \code{\link{Clarity_fixedK}}
}
\description{
Given data Y, and its regular svd Ysvd, and a choice of the number of SVD dimensions k to use.
}
\seealso{
\code{\link{Clarity_fixedK}} for the recommended interface.
}
