% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClaritySVD.R
\name{c_SimplexSurrounding}
\alias{c_SimplexSurrounding}
\title{Fit a simplex around data}
\usage{
c_SimplexSurrounding(x)
}
\arguments{
\item{x}{dataset to be enclosed, with data examples in rows and features in columns}
}
\value{
A matrix with each vertex as a column, of dimenison k+1
}
\description{
construct a simplex around the data x.
An arbitrary orientation is chosen.
}
\seealso{
\code{\link{c_scaleSimplex}} is used to compute the simplex
}
