% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityCompare.R
\name{c_SplitComparison}
\alias{c_SplitComparison}
\title{Perform a comparison on a list of resampled matrices}
\usage{
c_SplitComparison(
  Ydata1,
  Ydata2 = NULL,
  Yalt = NULL,
  transform = c_Procrust,
  k = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{Ydata1}{Either a list of matrix triples, or Y1'', the matrix to use to construct the Clarity model}

\item{Ydata2}{(default=NULL): If NULL, we assume Ydata1 is a list of matrix triples. If non-null, the matrix Y1', i.e. replicated data from the same distribution as Ydata1}

\item{Yalt}{(default=NULL): Should be NULL if Ydata2 is NULL. If non-null, should be Y2', that is a sampled alternative (dis)similarity matrix to be queried.}

\item{transform}{(default=\code{link{c_Procrust}}) a function taking two values, the first of which is transformed into the second. Use =\code{link{c_Identity}} to disable.}

\item{k}{(default=NULL) If null, perform \code{\link{Clarity_Scan}} prediction for Persistences. Otherwise perform \code{\link{Clarity_fixedK}} at the specified k.}

\item{verbose}{(default=FALSE) verbosity if operating over a list of matrix triples.}

\item{...}{Additional parameters to Clarity. The kmax parameter for  \code{\link{Clarity_Scan}} could be important.}
}
\value{
If provided with a list of matrix triples, return a list of matrix pairs giving the summary of the Clarity model predicting [Ydata2,Yalt]. If provided with three matrices, return two matrix predicting [Ydata2,Yalt]. For fixed k, this is a matrix of Residuals from \code{\link{Clarity_Extract}}; otherwise a matrix of Persistences from \code{\link{Clarity_Persistence}}.
}
\description{
Given a list of [Y1'', Y1', Y2'] matrix triples, learn a Clarity model on Y1'' , and predict Y1' and Y2' from that model. Do this for each triple and report the results as a pair of distributions, that for Y1' and for Y2'.
}
\seealso{
\code{\link{Clarity_Compare}}, the recommended way to access this code.
}
