% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityCompare.R
\name{c_SplitData}
\alias{c_SplitData}
\title{Split two datasets into three parts by downsampling}
\usage{
c_SplitData(D1, D2, distfn = c_dist, nbs = 100)
}
\arguments{
\item{D1}{reference matrix consisting N items with L1 features}

\item{D2}{reference matrix consisting N items with L2 features}

\item{distfn}{(default: \code{\link{c_dist}}) summary to apply, which should return a matrix of size N times N}

\item{nbs}{(default: 100) Number of times to repeat this process (set to NULL to do it only once and return a single resample list  [Y1'',Y1',Y2']}
}
\value{
A list of length nbs, each of which is a list containing [Y1'',Y1',Y2']
}
\description{
Downsample columns of D1 and D2 into three components:
  D1'': Random half of the columns (rounded up) of D1
  D1' : The other half of the columns (rounded down) of D1
  D2' : Random half of the columns (rounded down) of D2
distfn is evaluated on each to return (dis)similarities.
}
\seealso{
\code{\link{Clarity_Compare}}, the recommended way to access this code.
}
