% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClaritySVD.R
\name{c_svd}
\alias{c_svd}
\title{Compute mean centered svd}
\usage{
c_svd(Y)
}
\arguments{
\item{Y}{matrix to compute centered svd for}
}
\value{
the svd of Y as returned by base::svd
}
\description{
A shorthand for
    \code{svd(Y-rowMeans(Y))}
}
