% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClaritySimpleSVD.R
\name{c_svd0}
\alias{c_svd0}
\title{Compute the SVD of Y in convenient representation for CLARITY}
\usage{
c_svd0(Y)
}
\arguments{
\item{Y}{matrix to be fit}
}
\value{
A list containing u, d, and v as returned by svd.
}
\description{
Compute the SVD of Y, but unlike in core SVD, we use the convention that u[,i] and v[,i]
are positively correlated, and d can be be negative. This makes it more like the PCA represetation
and u==v if Y was symmetric.
}
