% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClaritySVD.R
\name{c_SVD_fixedK}
\alias{c_SVD_fixedK}
\title{Compute the Clarity model by SVD embedding data for a fixed size of model k}
\usage{
c_SVD_fixedK(Ysvd, Y, k, verbose = TRUE)
}
\arguments{
\item{Ysvd}{SVD of Y-rowMeans(Y)}

\item{Y}{matrix to be fit}

\item{k}{the number of mixture components to use}

\item{verbose}{(default TRUE) whether to output progress to the terminal.}
}
\value{
A Clarity object as described in \code{\link{Clarity_fixedK}}, with the additional elements:
\itemize{
\item Z The SVD locations of the data in k-1 dimensional space
\item ZS The k corners of the k-1 dimensional simplex enclosing the data
}
}
\description{
Given data Y, and its mean centred svd Ysvd, and a choice of the number of mixture components k to use,.
}
\seealso{
\code{\link{Clarity_fixedK}} for the recommended interface.
}
