% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClaritySVD.R
\name{c_SVD_ResidualExtend}
\alias{c_SVD_ResidualExtend}
\title{Extend a Clarity result by fitting to the residuals, using the SVD method and a range of k}
\usage{
c_SVD_ResidualExtend(clist, kmax = 10, verbose = TRUE)
}
\arguments{
\item{clist}{the Clarity object that was learned, as return by \code{\link{Clarity_Predict}}}

\item{kmax}{(Default: 10) The maximum number of additional mixture components to be considered}

\item{verbose}{(default TRUE) Whether to output progress information}
}
\value{
A ClarityScanExtend object as described in \code{\link{Clarity_Extend}}
}
\description{
Given a Clarity object, and details of the residuals that it has failed to fit, fit an additional 0..kmax mixture components to it to learn its unique model structure.  This uses the spectral embedding of the original fit, and the spectral embedding of the residuals, to make a feature vector that can be separated into the two components.
}
\seealso{
\code{\link{Clarity_Extend}} for the recommended interface.
}
