% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClaritySVD.R
\name{c_SVD_Scan}
\alias{c_SVD_Scan}
\title{Compute the Clarity model by SVD embedding data for a range of model sizes}
\usage{
c_SVD_Scan(Y, kmax = NULL, Ysvd = NULL, verbose = TRUE)
}
\arguments{
\item{Y}{matrix to be fit}

\item{kmax}{(default: NULL, meaning use the data dimension) the maximum number of mixture components to use}

\item{Ysvd}{(default: NULL, meaning calculate it) the SVD of Y-rowMeans(Y). If you provide a different object containing locations in the list Ysvd$u, they will be used instead.}

\item{verbose}{(default TRUE) whether to output progress to the terminal.}
}
\value{
A ClarityScan object as described in \code{\link{Clarity_Scan}}, with the additional elements:
\itemize{
\item Ysvd The SVD that was used in the embedding.
}
}
\description{
Given data Y, embed it into a space of dimension 1..kmax and fit it as a mixture
}
\seealso{
\code{\link{Clarity_Scan}} for the recommended interface.
}
