% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClaritySimpleSVD.R
\name{c_svdlist}
\alias{c_svdlist}
\title{Compute a set of diagonal-corrected SVDs}
\usage{
c_svdlist(Y, kmax, Tmax = 25, tol = 1e-05, verbose = FALSE)
}
\arguments{
\item{Y}{matrix to be fit}

\item{kmax}{(default: NULL, meaning use the data dimension) the maximum number of mixture components to use}

\item{Tmax}{(default: 25) As passed to \code{\link{c_IterativeSVD}}.}

\item{tol}{(default: 1e-5) As passed to \code{\link{c_IterativeSVD}}.}

\item{verbose}{(default TRUE) whether to output progress to the terminal.}
}
\value{
A SVDlist object, which is a list of length kmax of SVDs as returned by \code{\link{c_IterativeSVD}}
}
\description{
Given data Y, embed it into a space of dimension 1..kmax and fit it as an SVD
}
\seealso{
\code{\link{Clarity_Scan}} for the recommended interface.
}
