% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityMultiplicative.R
\name{c_updateA}
\alias{c_updateA}
\title{Multplicative update for A in Clarity}
\usage{
c_updateA(A, X, Y, fixed = NULL, norm = TRUE)
}
\arguments{
\item{A}{An N by K non-negative matrix of mixture proportions}

\item{X}{A K by K non-negative matrix of latent weights}

\item{Y}{An N by N non-negative matrix of similarities}

\item{fixed}{IGNORE}

\item{norm}{Whether to return the updated A normalised to have row sum 1. Default: TRUE.}
}
\value{
A non-negative matrix of dimension N by K
}
\description{
This function improves the fit of

     Y = || A X A^T ||_2

by updating A. X and Y are assumed to be positive, and A is positive with rows summing to 1. A has dimension N by K.

Seeking to solve

            Y = A X t(A)

we minimise F = (1/2) || Y - A X t(A) ||_2

The calculation leads to

A_{t+1} = A [ t(Y) A X + Y A t(X) ]

            ---------------------------------

 [ A X t(A) A t(X) + A t(X) t(A) A X ]
}
\keyword{mixture}
