% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityMultiplicative.R
\name{c_updateX}
\alias{c_updateX}
\title{Safe multiplicative update for X in Clarity}
\usage{
c_updateX(A, X, Y)
}
\arguments{
\item{A}{An N by K non-negative matrix of mixture proportions}

\item{X}{A K by K non-negative matrix of latent weights}

\item{Y}{An N by N non-negative matrix of similarities}
}
\value{
A non-negative matrix of dimension K by K
}
\description{
This function improves the fit of
     Y = || A X A^T ||_2
by updating X using the inverse if possible but falling back to a multiplicative update if this fails. X and Y are assumed to be positive, and A is positive with rows summing to 1. A has dimension N by K.

We use \code{\link{c_updateX_solve}} where possible. However, sometimes X is of low rank and the calculation fails; we fall back on a mulitiplicative update in this case.

Seeking to solve 
            Y = A X t(A)
we minimise F = (1/2) || Y - A X t(A) ||_2 
The calculation leads to
X_{t+1} = X t(A) Y A / [ t(A) A X t(A) A ]
}
\seealso{
\code{\link{c_updateA}} for updates to A
}
\keyword{mixture}
