% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clarity.R
\name{c_updateX_solve}
\alias{c_updateX_solve}
\title{Exact update for X in Clarity}
\usage{
c_updateX_solve(A, Y)
}
\arguments{
\item{A}{A K by K non-negative matrix of latent weights}

\item{Y}{An N by N non-negative matrix of similarities}
}
\value{
A non-negative matrix of dimension K by K
}
\description{
This function improves the fit of
     Y = || A X A^T ||_2
by updating X. X and Y are assumed to be positive, and A is positive with rows summing to 1. A has dimension N by K.

Seeking to solve 
            Y = A X t(A)
            Y = A D t(A)
     t(A) Y A =  t(A) A D t(A) A
 G t(A) Y A G = G t(A) A D t(A) A G
              = D
where G = inverse t(A) A
}
\seealso{
c_updateX
}
\keyword{mixture}
