% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityBootstrap.R
\name{Clarity_Bootstrap}
\alias{Clarity_Bootstrap}
\title{Bootstrap resample data to obtain distribution of predictions under a specified Clarity model}
\usage{
Clarity_Bootstrap(D, clearned, nbs = 100, k = NULL, distfn = c_dist,
  summary = abs, diag = 0, seed = NULL, verbose = FALSE, ...)
}
\arguments{
\item{D}{an N by L matrix containing N data items and L features. The features are resampled with replacement.}

\item{clearned}{either a ClarityScan or Clarity object. If a Clarity object, then residuals will be bootstrapped. If a ClarityScan object then Persistence will be bootstrapped, unless a specific k is given in which case Residuals are again calculated.}

\item{nbs}{Number of bootstraps (default: 100)}

\item{k}{If specified and clearned is a ClarityScan object, residuals for this k are bootstrapped. (default: NULL, meaning use Persistences)}

\item{distfn}{Distance function to apply to the bootstrapped data, which should return a similarity or dissimilarity matrix. Default: c_dist, which is as,matrix(dist(x)).}

\item{summary}{(default=abs) Summarisation of the bootstrapped value when extracting Residuals. Default: abs, which allows residuals to be easily compared. Use summary=I to return the raw data. See \code{\link{Clarity_Extract}}.}

\item{diag}{(default=0) diagonal value to be applied when extracting Residuals. See \code{\link{Clarity_Extract}}.}

\item{seed}{Seed to reset the random number generator with. Default: NULL, meaning don't reset.}

\item{verbose}{Whether to print to screen each iteration. Default: FALSE}

\item{...}{Additional parameters to \code{\link{Clarity_Persistence}} (not required for normal usage)}
}
\value{
A list of length nbs, each containing a matrix
}
\description{
Takes a Clarity or ClarityScan object, and some feature data, and bootstrap resamples the features to result in a similarity matrix. This is predicted by the model and summarised appropriately for comparison to data. Uses \code{\link{Clarity_Persistence}} for Persistences and \code{\link{Clarity_Predict}} directly for Residuals.
}
