% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityBootstrap.R
\name{Clarity_BScompare}
\alias{Clarity_BScompare}
\title{Compare Observed to Bootstrap values}
\usage{
Clarity_BScompare(testbs, obs, population = "element",
  comparison = function(x, o) {     sum(c(x, o) >= o)/(length(x) + 1) })
}
\arguments{
\item{testbs}{A list of bootstrapped matrices as returned from \code{\link{Clarity_Bootsrap}}}

\item{obs}{A matrix of observed values}

\item{population}{(default="element") whether the population being compared to is the "element" X[i,j], or "bestcol": the B best values in the "column" X[,i]}

\item{comparison}{(default=function(x,o){sum(o>x)/(length(x)+1)}) a function to score the observed value relative to the bootstraped values.}
}
\value{
A matrix of the same shape as obs, containing the result of comparison applied to each element
}
\description{
For each element in an observed matrix, extract out the corresponding elements in a list of bootstrapped versions of that matrix. Report on the position of the observed value relative to the bootstrapped values. By default this returns the results in the form of a p-value for a test with the alternative that the observed value is larger than the bootstraps (and the null that it is not)

Note that when population="bestrow" a conservative test is applied in which the best B bootstrap values from the entire column is used as the null distribution. This is appropriate for testing Persistences extracted via \code{\link{Clarity_Persistence}}.
}
\seealso{
\code{\link{Clarity_Bootsrap}}
}
