% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityPlots.R
\name{Clarity_Chart}
\alias{Clarity_Chart}
\title{Plot a matrix in the Clarity "Persistence Chart" style}
\usage{
Clarity_Chart(x, signif = NULL, mar = c(6, 6, 4, 1), scalefun = sqrt,
  zlim = NULL, cols = (grDevices::colorRampPalette(c("white", "yellow",
  "orange", "red", "darkred")))(100), text = FALSE, cex.text = 1,
  digits = 2, axes = TRUE, las.axis = 1, cex.axis = 1,
  cex.axis.X = NULL, cex.axis.Y = NULL, axis.top = FALSE,
  etext = NULL, etextgap = 0.25, cex.etext = 1, userect = FALSE,
  rectdelta = c(0.8, 0.8), signiffade = "55", imageback = "white",
  tol = 1e-10, main = "", xlab = "", ylab = "", ...)
}
\arguments{
\item{x}{The matrix object to be drawn}

\item{signif}{(default=NULL) an optional logical matrix of the same dimension as x. If provided, instead of a regular image, entries that are FALSE are drawn smaller and faded. If not provided, the plot is drawn as if entries are TRUE.}

\item{mar}{(default=c(6,6,4,1) margins of the plot as in \code{\link{par}}(mar). Set to NULL to disable setting them.}

\item{scalefun}{(default=sqrt) scaling function for the entries. The default is appropriate when plotting squared values such as  square residuals, persistences, etc. function(x){log(1+x)} can often be a good choice, as can the identity I.}

\item{zlim}{(default=NULL) range of the colour scale. Defaults to the range of x. If zlim=="symmetric" then it is symmetric around 0.}

\item{cols}{(default=colorRampPalette(c("white","yellow","orange","red","darkred"))(100)) a colour scale. It is essential if signif is provided that the colors are provided in "#RRGGBBAA" for red/green/blue/alpha hex values.}

\item{text}{(default=FALSE) whether to write the values on each matrix element. This is useful for small matrices.}

\item{cex.text}{(default=1) text size if text==TRUE.}

\item{digits}{(default=2) number of digits to write matrix values to if text==TRUE.}

\item{axes}{(default=TRUE) whether to draw any axes.}

\item{las.axis}{(default=1) \code{\link{par}}(las) for axis text.}

\item{cex.axis}{(default=1) \code{\link{par}}(cex) for axis text.}

\item{cex.axis.X}{(default=NULL) cex for X axis. Inherits from cex.axis if NULL. Set to 0 to disable X axis.}

\item{cex.axis.Y}{(default=NULL) cex for Y axis. Inherits from cex.axis if NULL. Set to 0 to disable Y axis.}

\item{axis.top}{(default=FALSE) whether to draw the X axis at the top, rather than the bottom, of the image.}

\item{etext}{(default=NULL) an optional matrix of additional text to be drawn in each cell underneath the main text.}

\item{etextgap}{(default=0.25) distance below the center of each rectangle to place the etext content.}

\item{cex.etext}{(default=1) size of additional text.}

\item{userect}{(default=FALSE) whether to use the "rectangle" plotting style, or the image plotting style. This is overriddent to TRUE if signif is provided.}

\item{rectdelta}{(default=c(0.8,0.8)) scaling factor for (X,Y) for non-significant cells.}

\item{signiffade}{(default="55") fading in the alpha channel for non-significant cells.}

\item{imageback}{(default="white") for the rectangle plotting style, the background colour that will appear around the outside of non-significant cells.}

\item{tol}{(default=1e-10) for the rectangle plotting style, the tolerance for determining which colour bin entries should appear}

\item{main}{(default="") Title as in \code{\link{image}}.}

\item{xlab}{(default="") Title as in \code{\link{image}}.}

\item{ylab}{(default="") Title as in \code{\link{image}}.}

\item{...}{Additional parameters to \code{\link{image}}}
}
\value{
NULL, invisibly. This funciton is used for its side effects in plotting.
}
\description{
Takes a matrix and plots it as an advanced heatmap, including markup for significance, explicit scaling of colours, text on elements, and more.

NB The recommended interface is  \code{\link{plot.Clarity}} and \code{\link{plot.ClarityScan}}.
}
\examples{
\donttest{
scan=Clarity_Scan(dataraw)
predmix=Clarity_Predict(datamix,scan)
scanpred=Clarity_Predict(datamix,scan)
pvalmix=Clarity_Compare(scan,D=datarawD,Dnew=datamixD)

## The recommended way to plot:
plot(pvalmix)

## The manual way to plot:
## Extract observed persistences
P=Clarity_Persistence(predmix)
## Compute pvalues
pvals=pvalmix$pvals
## pvals is a matrix of dimension N by K
signif=pvals<0.01

## Now pass to Clarity_Chart
Clarity_Chart(P,signif)
}
}
\seealso{
\code{\link{plot.Clarity}}, \code{\link{plot.ClarityScan}} require much less manual effort to extract various matrices from Clarity and ClarityScan objects. \code{\link{c_legend}} for one way to present a legend.
}
