% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityCompare.R
\name{Clarity_Compare}
\alias{Clarity_Compare}
\title{Compare a new dissimilarity matrix to another via the learned Clarity model}
\usage{
Clarity_Compare(
  clearned,
  D = NULL,
  Dnew = NULL,
  H0 = "value",
  Ylist = NULL,
  Ynew = NULL,
  distfn = c_dist,
  checktol = sqrt(.Machine$double.eps),
  transform = NULL,
  comparison = c_CompareNormal,
  k = NULL,
  nbs = 100,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{clearned}{A Clarity or ClarityScan object, from which inference will be drawn. If a ClarityScan object and \code{k} is not set, persistences are examined; otherwise residuals are examined.}

\item{D}{(default=NULL) The *original* data feature matrix. If provided, the features will be resampled to generate replicated datasets and hence (with \code{distfn}) a list of resampled dis/similarity matrices \code{Ylist}.}

\item{Dnew}{(default=NULL) Optionally, the new data to be compared. Will be used to construct \code{Ynew} (with \code{distfn}) if that is not provided; otherwise it will be checked for consistency.}

\item{H0}{(default="value") The null hypothesis to be tested. (Acts only to set the \code{transform}, as specified in the "transform" parameter.}

\item{Ylist}{(default=NULL) A list of replicated dis/similarity matrices. Provide them this way if it is more convenient.}

\item{Ynew}{(default=NULL) The new (dis)similarity matrix to be compared. Can instead provide \code{Dnew} and \code{Ynew} will be computed.}

\item{distfn}{(default=c_dist) Distance function to apply to \code{Dnew} and replicates of \code{D} to produce \code{Ynew} and \code{Ylist}.}

\item{checktol}{(default=\code{sqrt(.Machine$double.eps)}) tolerance for checking that \code{Ynew} matches that computed from \code{Dnew}.}

\item{transform}{(default=NULL) Transformation to apply to \code{Ynew} and replicates \code{Ylist} to match them to \code{Y}. This is set by \code{H0} if provided, but can be one of \code{\link{c_Procrust}} (H0="structure") \code{\link{c_MeanScale}} (H0="scale") or \code{\link{c_Identity}} (H0="value")}

\item{comparison}{(default=\code{\link{c_CompareNormal}}) the comparison between observed and resampled values to compute p-values, controlling the alternative hypothesis H1. Choices are one-tailed tests (\code{\link{c_CompareNormal}}, \code{\link{c_CompareEmpirical}}) looking for large residuals/persistences, or two-tailed tests (\code{\link{c_CompareNormalTwoTailed}} and \code{\link{c_CompareEmpiricalTwoTailed}}) also looking for small values.}

\item{k}{(default=NULL) If \code{clearned} is a ClarityScan object, the Clarity object with this object is extracted and used instead.}

\item{nbs}{(defaul=100) Number of resamples to perform, if \code{D} is provided to construct replicate \code{Ylist}.}

\item{verbose}{(default TRUE) Verbosity level.}

\item{...}{Additional arguments to \code{\link{Clarity_Scan}} or \code{\link{Clarity_fixedK}}.}
}
\value{
If p-values are not computed, an object of the same class as \code{clearned} predicting the new data. Otherwise, a \code{ClaritySignificance} or \code{ClarityScanSignificance} object (if \code{clearned} was a Clarity or ClarityScan object respectively). This is a list consisting of:
\itemize{
\item Ynew: the new dis/similarity matrix
\item Yobs: the transformed new dis/similarity matrix
\item pred: the \code{Clarity}/\code{ClarityScan} object predicting \code{Yobs} from \code{clearned}
\item splitlist: a list of resampled results as returned from \code{\link{c_SplitComparison}}
\item pvals: the computed pvalues
\item H0: the provided H0
\item transform: the applied transform
\item distfn: the applied distfn
}
}
\description{
Compare a similarity matrix \code{Ynew} to the Clarity model \code{clearned} (learned on \code{Y}) by a) transforming Ynew to maximally look like Y; b) computing persistences or residuals; c) optionally, comparing these to replicated data that produce "alternative versions" of Y.

To skip p-value computation, only \code{Ynew} and \code{clearned} are required. In this case, the only difference from Clarity_Predict is that \code{Ynew} is first transformed with \code{transform} to "look like" the original data.

To compute p-values, provide additionally a) the original data feature matrix \code{Dnew} and its associated comparison function \code{distfn}, or b) a list of replicated Y (dis)similarity matrices \code{Ylist}. Each replicate must contain a list of three matrices: [Y0,Y1,Y2] where Y0 and Y1 are ferom disjoint samples of the original data, and Y2 is sampled from the new data Dnew similarly to Y1. See \code{\link{c_SplitData}} for how this is done on feature matrices.

Some additional checking is done to try to ensure that sane p-values can be computed.
}
\seealso{
\code{\link{Clarity_Scan}} for computing \code{clearned}; \code{\link{Clarity_Predict}} for prediction; \code{\link{Clarity_Compare}} for assessing uncertainty; \code{\link{plot.Clarity}} and \code{\link{plot.ClarityScan}}.
}
