% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clarity.R
\name{Clarity_ComparisonPlot}
\alias{Clarity_ComparisonPlot}
\title{Plot a Clarity object and another predicted from it}
\usage{
Clarity_ComparisonPlot(c1, c2, order = NULL, rownames = NULL,
  zlim1 = NULL, zlim2 = NULL, zlimresidual = NULL,
  name1 = "Learned", name2 = "Predicted", plotnames = c("Fit",
  "Data", "Residuals"), cex.axis = 1, cex.main = 1, range.cex = 0.5,
  range.line = 0, ...)
}
\arguments{
\item{c1}{The first, "learned" Clarity object}

\item{c2}{The second, "predicted" Clarity object}

\item{order}{(default NULL) the plotting order. If NULL, the data c1$Y are fit with a dendrogram to obtain a good ordering. If NA, no reordering is done. Otherwise a specified ordering can be given.}

\item{rownames}{(default NULL) the names to be given to the rows, in the original order. By default it is taken from rownames(c1$Y)}

\item{zlim1}{(default NULL) range of the prediction and data plots for c1. Default: range of those data.}

\item{zlim2}{(default NULL) as above for c1.}

\item{zlimresidual}{(default NULL) range of the residual plot. Default: range of residuals for c1 and c2. It is not allowed to use different scales for each residual.}

\item{name1}{(default "Learned") Name for the c1 data}

\item{name2}{(default "Predicted") Name for the c2 data}

\item{plotnames}{(default c("Fit","Data","Residuals")) names for the Prediction, Data, and Residuals plots. Set to "" to disable additional naming}

\item{cex.axis}{(default 1) cex for the row/column names. Set to 0 to disable axis plotting}

\item{cex.main}{(default 1) cex for the titles}

\item{range.cex}{(default 0.5) cex for the residual range (set to 0 to disable plotting)}

\item{range.line}{(default 0) line for the residual range text (for mtext)}

\item{...}{Extra arguments for image: consider using different palettes.}
}
\value{
A list containing objects from the above list of possible inputs.
}
\description{
Takes two Clarity objects over the same samples and plots both of predictions, their associated data, and their residuals.
}
\examples{
\donttest{
scanraw=Clarity_Scan(dataraw) # Generate an initial run
scanraw=Clarity_Scan(dataraw,clist=scanraw) # Run it for longer
# Apply it to a new dataset with the same structure
scanrepfromraw=Clarity_Predict(datarep,clist=scanraw) 
# Apply it to a new dataset with slightly different structure
scanmixfromraw=Clarity_Predict(datamix,clist=scanraw)
# Plot the case where the residuals shouldn't matter
Clarity_ComparisonPlot(scanraw$scan[[19]],scanrepfromraw$scan[[19]],
      name2="Predicted (no structural change)",zlimresidual=c(-1,1))
# Plot the case where the residuals should matter
Clarity_ComparisonPlot(scanraw$scan[[19]],scanmixfromraw$scan[[19]],
      name2="Predicted (one structural change)",zlimresidual=c(-1,1))
}
}
\seealso{
\code{\link{Clarity_Scan}} to generate an estimate for c1 to be used here, and \code{\link{Clarity_Predict}} to generate an estimate for c2.
}
\keyword{mixture}
