% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clarity.R
\name{Clarity_Extend}
\alias{Clarity_Extend}
\title{Extend a ClarityScan to match new data}
\usage{
Clarity_Extend(
  clist,
  kmax = 10,
  extend = "Residual",
  method = "SVD",
  verbose = TRUE
)
}
\arguments{
\item{clist}{A learned Clarity object, of class "Clarity".}

\item{kmax}{(default: 10): The number of additional components to add}

\item{extend}{(default: "Residual") extension method. Currently the only one available.}

\item{method}{(default: "SVD") representation method. Currently the only one available.}

\item{verbose}{(default: TRUE) verbose output?}
}
\value{
An object of class "ClarityScanExtend", which is a list containing the same items as class "ClarityScan" (see \code{\link{Clarity_Scan}}) with the following modifications:
\itemize{
\item scan: A list of length kmax, each containing an object of class "ClarityExtend" containing the same items as a "Clarity" object as returned from \item{\link{Clarity_fixedK}} with the following additional items:
\itemize{
\item k0: The initial k from which the extension was run
\item R: The distance matrix of the residuals, which was fitted in the extension
\item Rsvd: The SVD of the distance matrix of the residuals.
}
\item kmax: The provided kmax
\item Y: The original data to fit
\item R: The residual structure that was fit in the extension
\item Rsvd: The SVD of R
\item k0: the originally used k
\item klist: the dimensions of the objects in "scan"
\item objectives: the objectives of the scan
}
}
\description{
Extend a ClarityScan that has been fitted to new data, to additionally fit the residuals of that data.
First call \code{\link{Clarity_Predict}} to fit the object to the new data, then call this to fit the residuals.

This function is not yet ready for production.

Currently only \code{\link{c_SVD_ResidualExtend}} is available as a method.
}
\seealso{
\code{\link{Clarity_Scan}} to generate an estimate for clist to be used here, and \code{\link{Clarity_Predict}} to make an initial prediction of the new data.
}
\keyword{mixture}
