% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityBootstrap.R
\name{Clarity_Extract}
\alias{Clarity_Extract}
\title{Extract objects from a Clarity or ClarityScan object}
\usage{
Clarity_Extract(cscan, what = "Yresid", summary = abs, diag = 0,
  k = NULL)
}
\arguments{
\item{cscan}{A Clarity or ClarityScan object}

\item{what}{(default="Yresid") the object to extract}

\item{summary}{(default=abs) any summary to apply to the result. abs is useful to create residuals that rank naturally.}

\item{diag}{(default=0) What to do with the diagonal. If NULL, nothing is done; otherwise the diagonal is set to this value. The diagonal often has inflated residuals and hence should normally be set to 0 for comparison.}

\item{k}{(default=NULL) if specified and cscan is a ClarityScan object, only the value for the specified k is returned, allowing residuals for a single k to be extracted.}
}
\value{
A list containing what for each k in the ClarityScan
}
\description{
Extract objects such as the Yresid residuals from a ClarityScan object. This is useful for constructing Persistences or similar. Also provides a standardized interface to extract the absolute value of residuals.
}
\seealso{
\code{\link{Clarity_Persistence}} to which you can pass
}
