% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clarity.R
\name{Clarity_fixedK}
\alias{Clarity_fixedK}
\title{Run The full Clarity algorithm for a fixed number of mixtures}
\usage{
Clarity_fixedK(Y, k, method = "SVDX", verbose = TRUE, ...)
}
\arguments{
\item{Y}{An N by N non-negative matrix of similarities}

\item{k}{an integer giving the dimension of the fit to be used}

\item{method}{(Default: "SVDX") Method to use. Fuzzy matched to either "SVD","SVDX","SVDmix" or "Multiplicative"}

\item{verbose}{(Default TRUE) whether to print progress information.}

\item{...}{futher parameters to be passed to \code{\link{c_multiplicative_fixedK}} or \code{\link{c_SVD_fixedK}}}
}
\value{
A list of class "Clarity" containing the following:
\itemize{
\item Y, the provided non-negative N by N similarity matrix
\item A, the inferred non-negative N by K mixture matrix
\item X, the inferred non-negative K by K latent matrix
\item calX, the inferred non-negative K by K 'naturally scaled' latent matrix
\item C, the matrix C above relating X to calX
\item tCinv, the inverse of C^T, used in relating X to calX
\item objective, the objective function evaluated at the proposed solution
\item prediction, A X A^T
\item Yresid, the residuals at the proposed solution
\item k, the provided number of mixture components
\item ..., additional content returned from the method.
}
}
\description{
Learn a mixture model representation of a provided similarity matrix Y using Clarity. Either uses SVD ( \code{\link{c_SVD_fixedK}}) or Multiplicative method (\code{\link{c_multiplicative_fixedK}}); see \code{\link{Clarity_Scan}} for a high level overview.
}
\examples{
\dontrun{
scansvd=Clarity_fixedK(dataraw,3) # Use the SVD method
scanmult=Clarity_fixedK(dataraw,3,method="M") # Use the Multiplicative method
}
}
\seealso{
The methods are \code{\link{c_SVD_fixedK}}) for SVD or \code{\link{c_multiplicative_fixedK}} for Multiplicative. \code{\link{Clarity_Scan}} is the recomended interface and gives a high level overview.
}
\keyword{mixture}
