% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityMultiplicative.R
\name{Clarity_Multiplicative_Scan}
\alias{Clarity_Multiplicative_Scan}
\title{Run The full Clarity algorithm}
\usage{
Clarity_Multiplicative_Scan(Y, kmax = 20, clist = NULL, niter = 10,
  thresh = 0.1, tmax = 10000, verbose = TRUE,
  matrixdistmin = 1e-08, objectivedistmin = 0.001, ...)
}
\arguments{
\item{Y}{An N by N non-negative matrix of similarities}

\item{kmax}{(default 20) an integer giving the maximum dimension of the fit to be used (2..kmax are used). Ignored if clist is provided.}

\item{clist}{(default NULL) A list of Clarity objects as returned by \code{\link{Clarity_Scan}}. If not provided, \code{\link{c_ScanOnly}} is used to generate an initial run with the specified kmax.}

\item{niter}{(default 10) maximum number of forward-backward iterations}

\item{thresh}{(default 0.1) required improvement in objective function across all K before we give up}

\item{tmax}{tmax as passed to \code{\link{Clarity_fixedK}}}

\item{verbose}{Whether to output information about progress. Note that Clarity_fixedK is always run with verbose=FALSE}

\item{matrixdistmin}{(default NULL meaning derive from clist) Passed to \code{\link{Clarity_fixedK}}.}

\item{objectivedistmin}{(default NULL meaning derive from clist) Passed to \code{\link{Clarity_fixedK}}.}

\item{...}{futher parameters to be passed to \code{\link{Clarity_fixedK}}}
}
\value{
An object of class "ClarityScan", which is a list containing the following:
\itemize{
\item scan A list of length (kmax)-1, each containing an object of class "Clarity" as described in \code{\link{Clarity_fixedK}} (for K=2..kmax)
\item objectives A vector of length (kmax)-1, the objective function at each K
\item klist The list of K evaluated at, K=2..kmax
\item Y The data Y provided
\item kmax The maximum K requested
}
}
\description{
Learn a mixture model representation of a provided similarity matrix Y using Clarity. You may either choose the maximum number of components to use, kmax, or provide a pre-existing list of
}
\examples{
\donttest{
scanraw=Clarity_Scan(dataraw) # Generate an initial run
scanraw=Clarity_Scan(dataraw,clist=scanraw) # Run it for longer
}
}
\seealso{
\code{\link{Clarity_fixedK}} is used at each K, and documents the important parameters. \code{\link{c_ScanOnly}} is used to generate the initial ClarityScan object independently for each K.
}
\keyword{mixture}
