% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityPlots.R
\name{Clarity_ObjectivePlot}
\alias{Clarity_ObjectivePlot}
\title{Plot the residuals of a set of ClarityScan objects as a function of k}
\usage{
Clarity_ObjectivePlot(clist, add = FALSE, legendloc = "bottomleft",
  legendbty = "n", name = NULL, lty = NULL, col = NULL,
  pch = NULL, lwd = NULL, lines = TRUE, points = TRUE,
  xlim = NULL, ylim = NULL, xlab = "K",
  ylab = "Objective Function", main = "", extern = TRUE, ...)
}
\arguments{
\item{clist}{the list of P clarityscan objects}

\item{add}{(default FALSE): if FALSE, call plot()}

\item{legendloc}{(default: "bottomleft") where the legend will be plotted. Set to non-chacter to disable legend plotting}

\item{legendbty}{(default: "n") box type (bty) in call to legend}

\item{name}{(Default: detect from clist, or use Scan1..P) names of clist items for legend}

\item{lty}{Line type. Default: 1}

\item{col}{Color. Default: 1..P}

\item{pch}{point type. Default: 1}

\item{lwd}{line width. Default: 1}

\item{lines}{(Default: TRUE) whether to plot lines}

\item{points}{(Default: TRUE) whether to plot points}

\item{xlim}{(Default: range of data) xlim for plot()}

\item{ylim}{(Default: range of data) ylim for plot()}

\item{xlab}{Label for x axis}

\item{ylab}{Label for y axis}

\item{main}{Label for main}

\item{extern}{(Default: TRUE) whether being called externally or not.}

\item{...}{Additional parameters to plot}
}
\value{
the desired plot symbol
}
\description{
Makes a plot of (objective function) against (k) for all ClarityScan objects, automatically doing sensible things with color, line type, legend, etc. For most arguments, try passing vectors of length P=number of plots to make.
}
