% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clarity.R
\name{Clarity_Predict}
\alias{Clarity_Predict}
\title{Predict one Similarity from a Clarity or ClarityScan object}
\usage{
Clarity_Predict(Ynew, clist)
}
\arguments{
\item{Ynew}{The data to be predicted}

\item{clist}{The learned Clarity object, either of class "Clarity" or of class "ClarityScan"}
}
\value{
An object of the same class as clist provided, with updated Y, X and derived features.
}
\description{
Takes each Clarity object and learns a new X for the new data Ynew provided.
}
\examples{
\donttest{
scanraw=Clarity_Scan(dataraw) # Generate an initial run
scanraw=Clarity_Scan(dataraw,clist=scanraw) # Run it for longer
# Apply it to a new dataset with the same structure
scanrepfromraw=Clarity_Predict(datarep,clist=scanraw) 
# Apply it to a new dataset with slightly different structure
scanmixfromraw=Clarity_Predict(datamix,clist=scanraw) 
}
}
\seealso{
\code{\link{Clarity_Scan}} to generate an estimate for clist to be used here.
}
\keyword{mixture}
