% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityBootstrap.R
\name{Clarity_Procrust}
\alias{Clarity_Procrust}
\title{Perform a Procrustes transformation of one matrix into another}
\usage{
Clarity_Procrust(Ystart, Ytarget)
}
\arguments{
\item{Ystart}{the matrix to transform}

\item{Ytarget}{The target matrix}
}
\value{
A matrix which is the best fitting transformation of Ystart to Ytarget
}
\description{
Procrust the matrix provided in the first argument, into the target provided in the second.  Performs Translation and Dilation.
}
\seealso{
\code{\link{Clarity_Bootstrap}}, which uses this code.
}
