% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clarity.R
\name{Clarity_Scan}
\alias{Clarity_Scan}
\title{Run the full Clarity algorithm to fit a range of mixture models to data.}
\usage{
Clarity_Scan(Y, kmax = 20, method = "SVD", clist = NULL,
  verbose = TRUE, ...)
}
\arguments{
\item{Y}{The similarity or distance matrix to be fitted. In principle any square matrix is allowed but algorithmically you may have problems if it is non-negative or rank-deficient.}

\item{kmax}{(default: 20): The maximum number of mixture components to be considered. All values less than this are explored}

\item{method}{(Default: "SVD") Fuzzy matched to either "SVD" or "Multiplicative"}

\item{clist}{(Default: NULL, meaning start from scratch) A previously learned  "ClarityScan" object. Useful for adding additional iterations to \code{\link{Clarity_Multiplicative_Scan}}.}

\item{verbose}{(default: TRUE) verbose output?}

\item{...}{additional parameters to the methods; principally tuning parameters for  \code{\link{Clarity_Multiplicative_Scan}}.}
}
\value{
An object of class "ClarityScan", which is a list containing:
\itemize{
\item scan A list of length (kmax)-1, each containing an object of class "Clarity" as described in \code{\link{Clarity_fixedK}} (for K=1..kmax)
\item objectives A vector of length (kmax), the objective function at each K
\item klist The list of K evaluated at, K=1..kmax
\item Y The data Y provided
\item kmax The maximum K requested
\item ... Additional method-specific details (e.g. Ysvd for method="SVD")
}
}
\description{
Learn a mixture model representation of a provided similarity or distance matrix Y using Clarity.

Two methods are available:
"SVD": A fast method that provides high quality predictions and zero fuss. Uses \code{\link{c_SVD_Scan}}.
"Multiplicative": A slower method that tries to find a minimum volume encolosing simplex for the data, making the output into a legititate mixture solution with additional assumptions. Uses \code{\link{Clarity_Multiplicative_Scan}}.
}
\examples{
\donttest{
scansvd=Clarity_Scan(dataraw) # SVD method model fit
scanmult=Clarity_Scan(dataraw,method="M") # Multiplicative method model fit
scanmult=Clarity_Scan(dataraw,method="M",clist=scanmult) # Iterate the Multiplicative method longer
}
}
\seealso{
\code{\link{Clarity_fixedK}} for the details of what a "Clarity" object is. \code{\link{Clarity_Predict}} to make an initial prediction of the new data.
}
\keyword{mixture}
