% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_general_venn_diagram.R
\name{create.colors}
\alias{create.colors}
\title{Creating colors for venn.diagram}
\usage{
create.colors(colo)
}
\arguments{
\item{colo}{character vector containing the predefined colors for selected breeds or approaches}
}
\value{
The function returns a character vector containing all colors for Venn diagram to be plotted
}
\description{
The function creates the colors for the intersection subsets for the Venn diagram to be plotted based on the predefined colors for approaches or breeds.
Venn diagram colors are possible to create for two or three sets.
}
\details{
function
}
\seealso{
* \link{mod_hotspot_server} \cr
* \link{mod_bc_general_server} \cr
* \link{mod_bc_genetic_map_server} \cr
* \link{mod_bc_hotspot_server}
}
