% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_general_fct_functions.R
\name{create.output}
\alias{create.output}
\title{Creating the correlation output for selected approaches (breed analysis)}
\usage{
create.output(data.trans, approach.info)
}
\arguments{
\item{data.trans}{data frame \link{data_plot_general}}

\item{approach.info}{data frame containing the predefined settings and names for the selected approaches (\link{table_approach_information})}
}
\value{
The function returns a character string.
}
\description{
The Pearson correlations between physical and genetic length are calculated for all selected approaches and summarized in a common output.
}
\details{
function
}
\seealso{
\link{mod_general_server}
}
