% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_general_venn_diagram.R
\name{creating_venn}
\alias{creating_venn}
\title{Creating the Venn diagram plot}
\usage{
creating_venn(venn_data, long.names, abbreviations, venn.colors)
}
\arguments{
\item{venn_data}{object of class "Venn" processed data (containing the plot data; created in \link{process_venn_data})}

\item{long.names}{vector containing the full names of selected breeds or approaches}

\item{abbreviations}{vector containing the predefined abbreviations for selected breeds or approaches}

\item{venn.colors}{vector containing the colors for the Venn sets (\link{create.colors})}
}
\value{
The function returns the plotted Venn diagram (ggplot object).
}
\description{
This function is used to create the legend text and the text used for each Venn set. The Venn diagram plot is then created.
}
\details{
function
}
\seealso{
* \link{mod_bc_general_server} \cr
* \link{mod_bc_genetic_map_server} \cr
* \link{mod_bc_hotspot_server} \cr
* \link{mod_hotspot_server}
}
