% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-data-description.R
\docType{data}
\name{data_plot_general}
\alias{data_plot_general}
\title{Data}
\description{
Data is described as required for the \link{scatterPlot_general}.
Moreover this also forms the basis for the correlation output (see \link{create.output}).
}
\details{
\tabular{rlll}{
   Column.No \tab Column.Name \tab Meaning \tab Class \cr
   1 \tab Mbp \tab Chromosome length in megabase pair \tab numeric \cr
   2 \tab cM \tab Chromosome length in centiMorgan for approach/breed \tab numeric \cr
   3 \tab Chr \tab Chromosome/s \tab character \cr
   4 \tab Color \tab Color for approaches/breeds \tab character \cr
   5 \tab What \tab Names of the approaches or breeds \tab character \cr
   6 \tab Size \tab Scatter plot size \tab numeric \cr
}
}
\seealso{
\itemize{
\item created in \link{transformdata_general_bc} or \link{transformdata_general} \cr
\item used in \link{scatterPlot_general} and \link{create.output}
}
}
\keyword{data}
