% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-data-description.R
\docType{data}
\name{data_plot_geneticMap}
\alias{data_plot_geneticMap}
\title{Data}
\description{
Data is described as required for \link{makePlot_geneticMaps} and \link{makePlot_all_geneticMaps}.
}
\details{
\tabular{rlll}{
   Column.No \tab Column.Name \tab Meaning \tab Class \cr
   1 \tab Chr \tab Chromosome \tab integer \cr
   2 \tab X \tab Position of SNP marker based on physical map in megabase pair \tab numeric \cr
   3 \tab Y \tab Position of SNP marker based on genetic map (approach) in centiMorgan \tab numeric \cr
   4 \tab Marker \tab SNP marker name \tab character \cr
   5 \tab Legend \tab Names of the selected approaches or breeds \tab character \cr
   6 \tab Color \tab Color for selected approaches or breeds \tab character \cr
}
}
\seealso{
\itemize{
\item created in \link{transformdata_general_bc} or \link{transformdata_general} \cr
\item used in \link{makePlot_geneticMaps} and \link{makePlot_all_geneticMaps}
}
}
\keyword{data}
