% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-data-description.R
\docType{data}
\name{data_plot_hotspot}
\alias{data_plot_hotspot}
\title{Data}
\description{
Data is described as required for the \link{scatterPlot_hotspot}.
}
\details{
\tabular{rlll}{
   Column.No \tab Column.Name \tab Meaning \tab Class \cr
   1 \tab Chr \tab Chromosome \tab integer \cr
   2 \tab SNP \tab SNP marker name \tab character \cr
   3 \tab BP \tab Marker position in base pair (physical map) \tab integer \cr
   4 \tab cM \tab Marker position based on selected approach/es (genetic map) \tab numeric \cr
   5 \tab Theta \tab Recombination rate \tab numeric \cr
   6 \tab Mbp \tab Marker position in base pair (physical map) \tab numeric \cr
   7 \tab what \tab Names of the approaches or breeds \tab character \cr
   8 \tab size \tab Size for each plot dot depending on hotspot or not (internal: 7-hotspot; 3-no hotspot) \tab numeric \cr
   9 \tab coloring \tab Color for each plot dot depending on hotspot or not \tab character \cr
   10 \tab shape \tab Shape for each plot dot depending on hotspot or not \tab numeric \cr
}
}
\seealso{
\itemize{
\item created in \link{transformdata_hotspot}  and \link{transformdata_hotspot_bc} \cr
\item used in \link{scatterPlot_hotspot}
}
}
\keyword{data}
