% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-data-description.R
\docType{data}
\name{data_table_geneticMap_bc}
\alias{data_table_geneticMap_bc}
\title{Data}
\description{
Data is described as required for the \link{prepare_table_venn}.
}
\details{
\tabular{llll}{
   Col.No \tab Column.Name \tab Meaning \tab Class \cr
   1 \tab Name \tab SNP marker name \tab character \cr
   2 \tab Chr \tab Chromosome number \tab integer \cr
   3 \tab Mbp_position \tab Marker position in megabase pair (physical map) \tab numeric \cr
   4 \tab bp_position \tab Marker position in base pair (physical map) \tab numeric \cr
   5-n \tab *_cM: \tab Marker position based on selected breed/s (genetic map) \tab numeric \cr
}
}
\seealso{
\itemize{
\item created in \link{transformdata_general_bc} \cr
\item used in \link{prepare_table_venn}
}
}
\keyword{data}
