% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-data-description.R
\docType{data}
\name{data_table_hotspot}
\alias{data_table_hotspot}
\title{Data}
\description{
Data is described as required for the \link{prepare_table_venn}.
}
\details{
\tabular{llll}{
   Col.No \tab Column.Name \tab Meaning \tab Class \cr
   1 \tab Name \tab SNP marker name \tab character \cr
   2 \tab Chr \tab Chromosome number \tab integer \cr
   3 \tab bp_position \tab Marker position in base pair (physical map) \tab numeric \cr
   4-n \tab *_cM \tab Marker position based on selected approach/es (genetic map) - breed analysis \tab numeric \cr
    \tab  \tab or Marker position based on selected breed/s (genetic map) -breed comparison \tab  \cr
}
}
\seealso{
\itemize{
\item created in \link{transformdata_hotspot} or \link{transformdata_hotspot_bc} \cr
\item used \link{prepare_table_venn}
}
}
\keyword{data}
