% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-data-description.R
\docType{data}
\name{data_traffic_light}
\alias{data_traffic_light}
\title{Data}
\description{
Data is described as required for the \link{make_traffic_light}.
}
\details{
\tabular{llll}{
   Column.No \tab Column.Name \tab Meaning \tab Class \cr
   1 \tab Breed \tab  \tab character \cr
   2 \tab n1 \tab Number of genotyped animals \tab integer \cr
   3 \tab N1 \tab Number of half-sib families \tab integer \cr
   4 \tab n2 \tab Number of genotyped animals in half-sib families with at least 30 progeny \tab integer \cr
   5 \tab N2 \tab Half-sib families with sires having at least 30 progenies \tab integer \cr
   6 \tab p \tab Number of SNPs \tab integer \cr
   7 \tab N2.new \tab Contains the value of N2 rounded to the nearest multiple 100 \tab numeric \cr
   8 \tab Averag.n2 \tab Contains the ratio of n2 to N2 to the nearest multiple of 100. \tab numeric \cr
}
}
\seealso{
\link{make_traffic_light}
}
\keyword{data}
