% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datamix}
\alias{datamix}
\title{Simulated Similarity dataset with modified topology for use in Clarity}
\format{A non-negative similarity matrix of dimension 100 by 100}
\source{
Uses Package \code{\link[ClaritySim]{ClaritySim}}
}
\usage{
datamix
}
\description{
A dataset containing the similarities generated by changing branch lengths of \code{\link{dataraw}} using \code{\link[ClaritySim]{mixCoalescent}} from package ClaritySim
}
\keyword{datasets}
