% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datamixD}
\alias{datamixD}
\title{Simulated Similarity dataset for use in Clarity}
\format{
A non-negative features matrix of dimension 100 by 500
}
\source{
Uses Package \code{\link[ClaritySim]{ClaritySim}}
}
\usage{
datamixD
}
\description{
A matrix containing the features for the generated by the following:

\code{library(ClaritySim)}
\code{set.seed(1)}
# Simlulate 100 objects in 10 hierarchically related clusters with 500 features
\code{mysim=simulateCoalescent(N=100,K=10,L=500,
                            alpha=0,
                            sigma0=0.0001,Amodel="uniform",
                            minedge=0.1)}
\code{mixsim<-mixCoalescent(mysim)}
\code{datamixD=mixsim$D}
}
\seealso{
Similarity matrix \code{\link{datamix}}, Similarity matrix with same topology \code{\link{dataraw}}, Similarity matrix with different topology \code{\link{datamix}}.
}
\keyword{datasets}
