% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataraw}
\alias{dataraw}
\title{Simulated Similarity dataset for use in Clarity}
\format{A non-negative similarity matrix of dimension 100 by 100}
\source{
Uses Package \code{\link[ClaritySim]{ClaritySim}}
}
\usage{
dataraw
}
\description{
A dataset containing the similarities generated by the following:
}
\details{
\code{set.seed(1)}
\code{mysim<-simulateCoalescent(100,10,rep(0.2,10))} # Simlulate 100 objects in a 10 dimensional latent space
\code{myroworder=order(apply(mysim$A,1,which.max))}
\code{mysim$A=mysim$A[myroworder,]}
mysim$Y=mysim$Y[myroworder,myroworder]
mysim$Y0=mysim$Y0[myroworder,myroworder]
dataraw=mysim$Y
}
\keyword{datasets}
