% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datarep}
\alias{datarep}
\title{Simulated Similarity dataset with modified branch lengths for use in Clarity}
\format{A non-negative similarity matrix of dimension 100 by 100}
\source{
Uses Package \code{\link[ClaritySim]{ClaritySim}}
}
\usage{
datarep
}
\description{
A dataset containing the similarities generated by changing branch lengths of \code{\link{dataraw}} using \code{\link[ClaritySim]{transformCoalescent}} from package \code{\link[ClaritySim]{ClaritySim}}
}
\keyword{datasets}
