% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_genetic_map_fct_functions.R
\name{get_only_legend}
\alias{get_only_legend}
\title{Extracting legend}
\usage{
get_only_legend(plot)
}
\arguments{
\item{plot}{ggplot object}
}
\value{
The function return the legend from a ggplot object.
}
\description{
The function extracts the legend of a ggplot object.
}
\details{
function
}
\note{
adopted from https://www.geeksforgeeks.org/add-common-legend-to-combined-ggplot2-plots-in-r/
}
\seealso{
The function is called from \link{makePlot_all_geneticMaps_bc}.
}
