% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_general_hovering.R
\name{hovering}
\alias{hovering}
\title{Creating hover information}
\usage{
hovering(dat1, hover, what)
}
\arguments{
\item{dat1}{data frame which was used to create the plot}

\item{hover}{list containing information from the scatter plot to show the corresponding point information
when mouse is moved over the plot.}

\item{what}{numeric specifies the outcome as well as the hovering style for the wellpanel where
 \enumerate{
   \item represents the \link{mod_general_server}
   \item represents the \link{mod_hotspot_server}
   \item represents the \link{mod_bc_general_server}
   \item represents the \link{mod_bc_hotspot_server}
}}
}
\value{
The function returns the hover information with corresponding style format.
}
\description{
The hover function determines the nearest point and creates style of the hover outcome.
}
\details{
function
}
\note{
The code was adopted from https://gitlab.com/-/snippets/16220 (Pawel github)
}
\seealso{
\link{mod_general_server}, \link{mod_hotspot_server}, \link{mod_bc_general_server}, and \link{mod_bc_hotspot_server}
}
