% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_genetic_function_fct_functions.R
\name{make_barplot_bc}
\alias{make_barplot_bc}
\title{Create barplot for genetic-map functions for selected breeds}
\usage{
make_barplot_bc(breed.infos, counts)
}
\arguments{
\item{breed.infos}{data frame containing the predefined settings and names for the selected breeds (\link{table_breed_information})}

\item{counts}{numeric containing for each selected breed the determined frequencies for each genetic-map function}
}
\value{
The function returns a ggplot object.
}
\description{
The function combines counts and breed information together, whereby the order based on genetic-map function.
}
\details{
function
}
\seealso{
\link{mod_bc_genetic_function_server}
}
