% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_genetic_map_fct_functions.R
\name{makePlot_all_geneticMaps}
\alias{makePlot_all_geneticMaps}
\title{Creating genetic map plots for all chromosomes}
\usage{
makePlot_all_geneticMaps(ll.gm.s)
}
\arguments{
\item{ll.gm.s}{list each list element contains a data frame (\link{data_plot_geneticMap}) for a chromosome for selected approaches (breed analysis) or the selected breeds (breed comparison)}
}
\value{
The function returns a patchwork object (containing multiple ggplot objects)).
}
\description{
The function creates the plots for all chromosomes and arranges them together in a single plot.
}
\details{
function
}
\seealso{
\link{mod_genetic_map_server}
}
