% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_genetic_map_fct_functions.R
\name{makePlot_all_geneticMaps_bc}
\alias{makePlot_all_geneticMaps_bc}
\title{Creating all genetic map plots for all chromosomes for more than one breed}
\usage{
makePlot_all_geneticMaps_bc(ll, names.bc.plot, colo)
}
\arguments{
\item{ll}{list whereby each list element contains a data frame including chromosome number, the position in Mbp and cM, as well as breed information for a chromosome}

\item{names.bc.plot}{vector containing names of selected breeds}

\item{colo}{vector containing the pre-defined colors}
}
\value{
The function return a grid with arranged ggplots.
}
\description{
The function creates the plots for all chromosomes for more than one breed and grid arrange them. To obtain only one
legend for all plots, the function \link{get_only_legend} is called inside.
}
\details{
function
}
\note{
adopted from https://www.geeksforgeeks.org/add-common-legend-to-combined-ggplot2-plots-in-r/
}
\seealso{
\link{mod_bc_genetic_map_server}
}
