% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_genetic_function_fct_functions.R
\name{makePlot_genetic_function}
\alias{makePlot_genetic_function}
\title{Creating plots of genetic-map functions for breed analysis}
\usage{
makePlot_genetic_function(
  chromo = chromo,
  df.list = df.list,
  name.file,
  breed.select
)
}
\arguments{
\item{chromo}{selected chromosome}

\item{df.list}{list: first element contains a matrix of recombination rate and distance for marker pairs, second element includes a matrix of X-values for all genetic-map functions,
third element includes a matrix of Y-values for all genetic-map functions.}

\item{name.file}{name of output file.}

\item{breed.select}{selected breed}
}
\value{
The functions returns an already rendered plotly object.
}
\description{
For a selected chromosome the scatter plot of recombination rates vs. distances and curves of all genetic-map functions ("Haldane scaled","Rao","Felsenstein","Liberman & Karlin")
are generated.
}
\details{
function

The plots are cached to improve the app performance.
}
\seealso{
\link{mod_genetic_function_server}
}
