% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_genetic_map_fct_functions.R
\name{makePlot_geneticMaps}
\alias{makePlot_geneticMaps}
\title{Creating the genetic map plot for breed analysis}
\usage{
makePlot_geneticMaps(chr, dat_maps, name.file, colo)
}
\arguments{
\item{chr}{character contains the selected chromosome}

\item{dat_maps}{data frame containing the genetic map data}

\item{name.file}{character contains the name for image file}

\item{colo}{vector contains the predefined colors}
}
\value{
The function returns a plotly object.
}
\description{
The function creates the plot for a single chromosome for one breed.
}
\details{
function
}
\seealso{
\link{mod_genetic_map_server}
}
