% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_genetic_map_fct_functions.R
\name{makePlot_geneticMaps_bc}
\alias{makePlot_geneticMaps_bc}
\title{Creating the genetic map plot for breed comparison}
\usage{
makePlot_geneticMaps_bc(chr, dat_maps.list, colo, plot.pos, what, name.file)
}
\arguments{
\item{chr}{selected chromosome}

\item{dat_maps.list}{list each list element contains the genetic map data for a breed}

\item{colo}{vector containing the pre-defined colors}

\item{plot.pos}{contains the column which should be used}

\item{what}{contains the approach name ("deterministic" or "likelihood-based") to create plot title}

\item{name.file}{name of output file}
}
\value{
The function return a plotly object.
}
\description{
The function creates the plot for a single chromosome for more than one breed.
}
\details{
function
}
\seealso{
\link{mod_bc_genetic_map_server}
}
