% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bc_general.R
\name{mod_bc_general_ui}
\alias{mod_bc_general_ui}
\alias{mod_bc_general_server}
\title{mod_bc_general_ui and mod_bc_general_server}
\usage{
mod_bc_general_ui(id)

mod_bc_general_server(
  id,
  filter,
  geneticMap.bc,
  names.files,
  make.traff.light.bc,
  breed.infos,
  approach,
  dt,
  figure_rend2_g
)
}
\arguments{
\item{id}{module id}

\item{filter}{character contains the selected chromosome}

\item{geneticMap.bc}{list containing the genetic maps for all selected breeds (each list element is a breed)}

\item{names.files}{character containing the concatenated breed names and is used as an addition to the file name}

\item{make.traff.light.bc}{object containing the ggplot for the likelihood quality signal (\link{make_traffic_light})}

\item{breed.infos}{data frame containing the predefined settings and names for the selected breeds (\link{table_breed_information})}

\item{approach}{character containing the abbreviation for the selected approach}

\item{dt}{list whereby each list element contains for each selected breed the genetic map summary information}

\item{figure_rend2_g}{reactive value indicates if the plot in the module has been fully rendered. Used on the main server to enable or disable "Breed comparison" sidebar elements.}
}
\description{
A shiny module to generate the outcome for the tabpanel "General information" for the sidebar "Breed comparison".
}
\details{
Within the module the data are prepared for the table and for plotting using the function \link{transformdata_general_bc} depending on user's selected chromosome.
The corresponding plots are created using the function \link{scatterPlot_general}.
 This module also shows a Venn diagram of corresponding markers. The corresponding input data are
 prepared using the function \link{process_venn_data} and plotted using the function \link{creating_venn}.
 Finally, the quality of the likelihood approach is also shown here by means of a traffic light (also termed as likelihood quality signal \link{make_traffic_light}).
}
\seealso{
* \link{transformdata_general_bc} \cr
* \link{scatterPlot_general} \cr
* \link{process_venn_data} \cr
* \link{creating_venn} \cr
* \link{make_traffic_light}
}
