% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bc_hotspot.R
\name{mod_bc_hotspot_ui}
\alias{mod_bc_hotspot_ui}
\alias{mod_bc_hotspot_server}
\title{mod_bc_hotspot_ui and mod_bc_hotspot_server}
\usage{
mod_bc_hotspot_ui(id)

mod_bc_hotspot_server(
  id,
  filter,
  breed.select.bc,
  color.shape.def,
  names.bc.venn,
  names.files
)
}
\arguments{
\item{id}{module id}

\item{filter}{selected chromosome}

\item{breed.select.bc}{vector containing the names of selected breeds}

\item{color.shape.def}{data frame containing the definition for coloring, shapes, size for plots}

\item{names.bc.venn}{vector containing the first the letter of the selected breed name}

\item{names.files}{string containing the concatenate breed names}
}
\description{
A shiny module to generate the outcome for the tabpanel "General information" for the sidebar "Breed comparison".
}
\details{
Within the module the input data are prepared using the function \link{transformdata_hotspot}. To produce the hotspot plot, the function
\link{scatterPlot_hotspot} or \link{scatterPlot_hotspot_all} depending on the user choice of chromosome is used. The hover information are obtained from the function \link{hovering}.
 This module also shows a Venn diagram of corresponding hotspot markers. The input data are prepared
 using the function \link{process_venn_data} and plotted using the function \link{creating_venn}. The user may select a Venn diagram subset and the
 table will be reduced accordingly using the function \link{prepare_table_venn}.
}
\seealso{
\link{transformdata_hotspot}, \link{scatterPlot_hotspot}, \link{scatterPlot_hotspot_all}, \link{hovering},\link{process_venn_data},\link{creating_venn} and \link{prepare_table_venn}
}
\keyword{internal}
