% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_contact.R
\name{mod_contact_ui}
\alias{mod_contact_ui}
\alias{mod_contact_server}
\title{mod_contact_ui and mod_contact_server}
\usage{
mod_contact_ui(id)

mod_contact_server(id)
}
\arguments{
\item{id}{module id}
}
\description{
A shiny module containing the contact information and references.
}
\keyword{data}
\keyword{internal}
\keyword{internal,}
\keyword{set}
