% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_datasets.R
\name{mod_datasets_ui}
\alias{mod_datasets_ui}
\alias{mod_datasets_server}
\title{mod_datasets_ui and mod_datasets_server}
\usage{
mod_datasets_ui(id)

mod_datasets_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module containing all relevant information about the data sets. Currently, only one Holstein data set is included.
}
\keyword{internal}
