% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_general.R
\name{mod_general_ui}
\alias{mod_general_ui}
\alias{mod_general_server}
\title{mod_general_ui and mod_general_server}
\usage{
mod_general_ui(id)

mod_general_server(
  id,
  filter,
  breed.select,
  make.traff.light,
  approach.info,
  dt,
  add.approach.names
)
}
\arguments{
\item{id}{module id}

\item{filter}{character contains the selected chromosome}

\item{breed.select}{character contains the name of the selected breed}

\item{make.traff.light}{object containing the ggplot for the likelihood quality signal (\link{make_traffic_light})}

\item{approach.info}{data frame containing the predefined settings and names for the selected approaches (\link{table_approach_information})}

\item{dt}{data frame contains the genetic map summary for the selected breed}

\item{add.approach.names}{character containing the concatenated approach abbreviations and is used as an addition to the file name}
}
\description{
A shiny module to generate the outcome for the tabpanel "General information" for the sidebar "Breed analysis".
}
\details{
Within the module the data are prepared for the table and for plotting using the function \link{transformdata_general}. Moreover,
the plots are created using the function \link{scatterPlot_general}.
The function \link{create.output} is used to create the outcome for the correlation depending on selected approaches.
Finally, if the likelihood-based approach is selected, the quality of the likelihood approach is also shown here by means of a traffic light
(also termed as likelihood quality signal (\link{make_traffic_light})).
}
\seealso{
* \link{transformdata_general} \cr
* \link{scatterPlot_general} \cr
* \link{make_traffic_light} \cr
* \link{create.output}
}
\keyword{general}
\keyword{internal,}
\keyword{module}
