% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_genetic_function.R
\name{mod_genetic_function_ui}
\alias{mod_genetic_function_ui}
\alias{mod_genetic_function_server}
\title{mod_genetic_function_ui and mod_genetic_function_server}
\usage{
mod_genetic_function_ui(id)

mod_genetic_function_server(id, filter, breed.select)
}
\arguments{
\item{id}{module id}

\item{filter}{selected chromosome}

\item{breed.select}{selected breed}
}
\description{
A shiny module to generate the outcome for the tabpanel "Genetic-map functions" for the sidebar "Breed analysis".
}
\details{
The module uses the function \link{makePlot_genetic_function} for plotting.
}
\seealso{
\link{makePlot_genetic_function}
}
\keyword{function}
\keyword{genetic}
\keyword{internal}
\keyword{internal,}
