% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_hotspot.R
\name{mod_hotspot_ui}
\alias{mod_hotspot_ui}
\alias{mod_hotspot_server}
\title{mod_hotspot_ui and mod_hotspot_server}
\usage{
mod_hotspot_ui(id)

mod_hotspot_server(input, output, session, val, filter)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module to generate the outcome for the tabpanel Hotspot detection
}
\note{
The server used functions which are included in mod-hotspot_fct_functions as well as the general fct_general_hovering function.
}
\keyword{internal}
