% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_methodology.R
\name{mod_methodology_ui}
\alias{mod_methodology_ui}
\alias{mod_methodology_server}
\title{mod_methodology_ui and mod_methodology_server}
\usage{
mod_methodology_ui(id)

mod_methodology_server(id)
}
\arguments{
\item{id}{module id}
}
\description{
A shiny module containing all relevant information about the methodology or the terminology of the methods mentioned within the app.
}
\keyword{internal,}
\keyword{methodology}
