% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_misplaced.R
\name{mod_misplaced_ui}
\alias{mod_misplaced_ui}
\alias{mod_misplaced_server}
\title{mod_misplaced_ui and mod_misplaced_server}
\usage{
mod_misplaced_ui(id)

mod_misplaced_server(input, output, session, filter)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module to generate the outcome for Misplaced markers.
}
\keyword{internal}
