% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_startscreen.R
\name{mod_startscreen_ui}
\alias{mod_startscreen_ui}
\alias{mod_startscreen_server}
\title{mod_startscreen_ui and mod_startscreen_server}
\usage{
mod_startscreen_ui(id)

mod_startscreen_server(id)
}
\arguments{
\item{id}{Internal parameters for shiny.}
}
\description{
A shiny Module to create the start screen.
}
\details{
startscreen UI Function
}
