% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityPlots.R
\name{plot.ClarityPlot}
\alias{plot.ClarityPlot}
\title{Plot a ClarityPlot object}
\usage{
\method{plot}{ClarityPlot}(x, order = NULL, ...)
}
\arguments{
\item{x}{A ClarityPlot object}

\item{order}{(default=NULL) preferred order. If NULL, use the provided ordering. Otherwise, either a numeric vector of length N, or a permutation of the rownames of the original similarity Y.}

\item{...}{Extra parameters to be passed to \code{\link{Clarity_Chart}}}
}
\value{
NULL, invisibly. This funciton is used for its side effects in plotting.
}
\description{
This is a convenience wrapper to \code{\link{Clarity_Chart}} for ClarityPlot objects returned from \code{\link{plot.Clarity}} or \code{\link{plot.ClarityScan}}. It will extract the Persistence P or Squared Residuals R, along with any signif matrix.
}
\examples{
\dontrun{
scan=Clarity_Scan(dataraw) ## Core Clarity
predmix=Clarity_Predict(datamix,scan) ## Core prediction
pvalmix=Clarity_Compare(scan,D=datarawD,Dnew=datamixD)

## Plotting: generate a plotting object
predplot=plot(predmix,plot=FALSE)
## Plot it
plot(predplot)
}
}
\seealso{
\code{\link{Clarity_Chart}} for the actual plotting, \code{\link{plot.Clarity}} for plotting Residuals, and \code{\link{plot.ClarityScan}} for plotting Persistences. \code{\link{c_legend}} for one way to present a legend.
}
