% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityPlots.R
\name{plot.ClarityScan}
\alias{plot.ClarityScan}
\title{Plot persistences from a ClarityScan object}
\usage{
\method{plot}{ClarityScan}(x, signif = NULL, A = NULL, order = NULL,
  rotate = FALSE, plot = TRUE, type = "mean", kmax = NULL,
  thresh = 0.01, diag = 0, f = "RowSumsSquared", what = "Yresid",
  summary = abs, population = "bestcol", verbose = FALSE, ...)
}
\arguments{
\item{x}{A ClarityScan object.}

\item{signif}{(default=NULL) either a ClarityScanBootstrap object as returned by \code{\link{Clarity_Bootstrap}}; or a matrix of p-values as returned by \code{\link{Clarity_BScompare}}; or a logical matrix of elements that are significant (the same dimensions as the similarities modelled in x). In all cases it is transformed to a logical matrix and passed to \code{\link{Clarity_Chart}}.}

\item{A}{(default=NULL) A mixture matrix by which to summarise the matrix using \code{\link{c_MergeSquare}}. This might e.g. represent a clustering of the elements in the similarities. It must be an N (number of subjects) by K' (number of desired clusters) matrix with named columns.}

\item{order}{(default=NULL) preferred order. If NULL, use the provided ordering. Otherwise, either a numeric vector of length N, or a permutation of the rownames of the original similarity Y.}

\item{rotate}{(default=FALSE) whether to rotate the Persistences. If FALSE (default) then the complexity K is the Y-axis. If TRUE then the complexity K is the X-axis.}

\item{plot}{(default=TRUE) whether to actually create the plot or store the computed matrices for later plotting with \code{\link{plot.ClarityPlot}}.}

\item{type}{(default="mean") if A is not null, should we report the "mean" or the "sum" of the elements in each pair of (soft) clusters?}

\item{kmax}{(default=NULL) if provided, persistences above kmax are not included in the plot.}

\item{thresh}{(default=0.01) threshold for translating bootstrapped p-values into significant/not-significant calls for the matrix signif.}

\item{diag}{(default=0) how to treat the diagonal of residuals when computing persistences, as used by \code{\link{Clarity_Persistence}}. Set to NULL to use the observed values, but the default 0 is recommended.}

\item{f}{(default="RowSumsSquared") how to summarise residuals into persistences, as used by \code{\link{Clarity_Persistence}}.}

\item{what}{(default="Yresid") what to extract from each complexity when computing persistences, as used by \code{\link{Clarity_Persistence}}.}

\item{summary}{(default=abs) how to summarise residuals into persistences, as used by \code{\link{Clarity_Persistence}}.}

\item{population}{(default="bestcol") how to define the null model when computing p-values from bootstrapped data using \code{\link{Clarity_BScompare}}. The default is conservative by comparing each element to the most extreme values in its row, but "element" can be used to compare each element to only bootstrapped versions of itself.}

\item{verbose}{(default=FALSE) whether to output information about what transformations are being done.}

\item{...}{Additional parameters to  \code{\link{Clarity_Chart}}.}
}
\value{
A ClarityPlot object (invisibly) which is a list containing:
\itemize{
\item clist The provided ClarityScan object
\item A The provided A matrix
\item P The extracted Persistence matrix, transformed and merged
\item P0 The raw Persistence matrix as returned from \code{\link{Clarity_Persistence}}
\item signif0 The raw logical significance matrix, either as provided or as returned from \code{\link{Clarity_BScompare}} if provided with a ClarityScanBootstrap object (same dimension as P0)
\item signif The significance matrix, transformed and merged (same dimension as P)
\item rotate logical indicating whether we rotated the matrices
}
}
\description{
Plot a Persistence Chart, which is an image representing which subjects are poorly explained by a Clarity fit.

This is the main plot fuction for persistences. It has several convenience manipulations to extract persistences and statistical significance associated with them via bootstrapping, as well as summarising them by clustering or mixtures.
}
\examples{
\donttest{
scan=Clarity_Scan(dataraw) ## Core Clarity
predmix=Clarity_Predict(datamix,scan) ## Core prediction
scanbootstrap=Clarity_Bootstrap(scan,
                                target=datamix,
                                D=datarawD)

## Plotting:
plot(predmix,signif=scanbootstrap)
## Merged clusters:
plot(predmix,signif=scanbootstrap,A=datarawA)
}
}
\seealso{
Uses \code{\link{Clarity_Chart}} for plotting. \code{\link{plot.ClarityPlot}} is called to perform that plotting; \code{\link{c_Merge}} which is used to perform simplification using the mixture matrix A. \code{\link{c_legend}} for one way to present a legend. \code{\link{Clarity_ObjectivePlot}} for plotting the objective function.
}
