% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityPlots.R
\name{plot.ClarityScanSignificance}
\alias{plot.ClarityScanSignificance}
\title{Plot residuals and p-values from a ClarityScanSignificance object}
\usage{
\method{plot}{ClarityScanSignificance}(x, ...)
}
\arguments{
\item{x}{A ClarityScanSignificance object as returned by \code{\link{Clarity_Compare}}}

\item{...}{Extra parameters to \code{\link{plot.ClarityScan}}. Of particular relevence is \code{thresh} which controls the significance threshold.}
}
\value{
A ClarityScanPlot object (see \code{\link{plot.ClarityScan}}.
}
\description{
Plot a Persistence Chart, which is an image representing which subjects are poorly explained by a Clarity fit.
}
