% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityPlots.R
\name{plot.ClaritySignificance}
\alias{plot.ClaritySignificance}
\title{Plot residuals and p-values from a ClaritySignificance object}
\usage{
\method{plot}{ClaritySignificance}(x, ...)
}
\arguments{
\item{x}{A ClaritySignificance object as returned by \code{\link{Clarity_Compare}}}

\item{...}{Extra parameters to \code{\link{plot.Clarity}}. Of particular relevence is \code{thresh} which controls the significance threshold.}
}
\value{
A ClarityPlot object (see \code{\link{plot.Clarity}}.
}
\description{
Plot a Residuals Chart, which is an image representing which subjects are poorly explained by a Clarity fit.
}
