% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_general_venn_diagram.R
\name{process_venn_data}
\alias{process_venn_data}
\title{Prepare data for Venn plot}
\usage{
process_venn_data(venn)
}
\arguments{
\item{venn}{object of class "Venn"}
}
\value{
The function returns the processed Venn data object (containing the plot data).
}
\description{
This function prepares the data for the Venn plot.
}
\details{
function
}
\seealso{
\link{mod_bc_general_server}, \link{mod_bc_genetic_map_server} and \link{mod_bc_hotspot_server}
}
