% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Random dirichlet RV}
\usage{
rdirichlet(n, alpha)
}
\arguments{
\item{n}{number of random dirichlet variables}

\item{alpha}{dirichlet parameter}
}
\value{
A matrix with each random variate in a row
}
\description{
Simulate n Dirichlet Random Variables with parameter alpha
}
