% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_general_fct_functions.R
\name{scatterPlot_general}
\alias{scatterPlot_general}
\title{Creating the general scatter plot}
\usage{
scatterPlot_general(dat, name.file)
}
\arguments{
\item{dat}{data frame containing the corresponding plot information (\link{data_plot_general})}

\item{name.file}{character containing the name used for the generated plot}
}
\value{
The functions returns a plotly object.
}
\description{
The corresponding scatter plot is created, when a specific chromosome or all chromosomes was selected within breed analysis or breed comparison.
}
\details{
function
}
\seealso{
* \link{mod_general_server} \cr
 * \link{mod_bc_general_server}
}
