% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_general_fct_functions.R
\name{scatterPlot_general_all}
\alias{scatterPlot_general_all}
\title{Creating the scatter plot for all chromosomes within breed analysis}
\usage{
scatterPlot_general_all(dat, colo)
}
\arguments{
\item{dat}{data frame containing basepairs (bp), centiMorgan (cM), chromosome (chr), color and approach}

\item{colo}{vector containing the predefined colors}
}
\value{
The function returns a ggplot object.
}
\description{
The scatter plot is created, when all chromosomes are selected within breed analysis.
}
\details{
function
}
\seealso{
\link{mod_general_server}
}
