% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_general_fct_functions.R
\name{scatterPlot_general_selected}
\alias{scatterPlot_general_selected}
\title{Creating the scatter plot for selected chromosome within breed analysis}
\usage{
scatterPlot_general_selected(dat, fil, colo)
}
\arguments{
\item{dat}{data frame containing basepairs (bp), centiMorgan (cM), chromosome (chr), color and approach}

\item{fil}{selected chromosome}

\item{colo}{vector containing the predefined colors}
}
\value{
The functions returns a ggplot object.
}
\description{
The corresponding scatter plot is created, when a specific chromosome was selected within breed analysis.
}
\details{
function
}
\seealso{
\link{mod_general_server}
}
