% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_general_fct_functions.R
\name{scatterPlot_general_selected_bc}
\alias{scatterPlot_general_selected_bc}
\title{Creating the scatter plot for selected chromosome for breed comparison}
\usage{
scatterPlot_general_selected_bc(dat, fil, names.bc.plot, colo)
}
\arguments{
\item{dat}{data frame containing basepairs (bp), centiMorgan (cM), chromosome (chr), breed and color}

\item{fil}{selected chromosome}

\item{names.bc.plot}{vector containing the names of selected breeds}

\item{colo}{vector containing the predefined colors}
}
\value{
The functions returns a ggplot object.
}
\description{
The corresponding scatter plot is created, when a specific chromosome is selected for the breed comparison.
}
\details{
function
}
\seealso{
\link{mod_bc_general_server}
}
