% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_hotspot_fct_functions.R
\name{scatterPlot_hotspot}
\alias{scatterPlot_hotspot}
\title{Creating the hotspot scatter plot for a specific chromosomome}
\usage{
scatterPlot_hotspot(dat1, fil, ranges, legend1)
}
\arguments{
\item{dat1}{data frame (containing: Chr, SNP, cM, BP, Theta, Dis, ord, size, coloring, and shape)}

\item{fil}{selected chromosome}

\item{ranges}{contains brush information}

\item{legend1}{data frame containing the legend information (contains: label, breaks, shape, and ord)}
}
\value{
The function returns a ggplot object.
}
\description{
The function creates the hotspot plot for a specific selected chromosome.
}
\details{
function

Note: legend1, the column breaks contains the color for the corresponding label, and column ord contains numbers used for sorting to keep the sequence (corresponds to ord in dat1).
}
\seealso{
\link{mod_hotspot_server} and \link{mod_bc_hotspot_server}
}
