% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-data-description.R
\docType{data}
\name{table_approach_information}
\alias{table_approach_information}
\title{Data}
\description{
\tabular{rlll}{
   Column.No \tab Column.Name \tab Meaning \tab Class \cr
   1 \tab Name \tab Name of the approach \tab character \cr
   2 \tab Abbreviation \tab Abbreviation of the approach \tab character \cr
   3 \tab Color \tab Color used for all plots \tab character \cr
   4 \tab Color2 \tab Color used only for the hotspot plot \tab character \cr
}
}
\details{
Table contains all predefined names and settings for all approaches to enable standardized outputs.
Table information is used in all app modules.
}
\seealso{
created in \link{app_server}
}
\keyword{data}
