% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_general_fct_functions.R
\name{transformdata}
\alias{transformdata}
\title{transformdata}
\usage{
transformdata(data1 = data1)
}
\arguments{
\item{data1}{contains data frame which should be transformed}
}
\value{
A dataframe.
}
\description{
The function transform the corresponding columns bp into Mbp and M into cM.
}
\details{
functions
}
