% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_genetic_map_fct_functions.R
\name{transformdata_genetic_map}
\alias{transformdata_genetic_map}
\title{Prepare genetic map data for plotting}
\usage{
transformdata_genetic_map(n, input)
}
\arguments{
\item{n}{number of chromosomes}

\item{input}{data frame containing the genetic map information}
}
\value{
The function returns a list containing the desired information for each chromosome.
}
\description{
The function prepares the genetic map information for a single breed.
}
\details{
function
}
\seealso{
\link{mod_genetic_map_server}
}
