% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_genetic_map_fct_functions.R
\name{transformdata_genetic_map_bc}
\alias{transformdata_genetic_map_bc}
\title{Prepare genetic map data for plotting for breed comparison}
\usage{
transformdata_genetic_map_bc(n, input, meth, names.bc.plot)
}
\arguments{
\item{n}{number of chromosomes}

\item{input}{list of genetic maps for the selected breeds}

\item{meth}{name of approach ("deterministic", "likelihood-based")}

\item{names.bc.plot}{contains the names of the selected breeds}
}
\value{
The function returns a list containing the desired information for all selected breeds for each chromosome.
}
\description{
The function prepares the genetic map information for more than one breed.
}
\details{
function
}
\seealso{
\link{mod_bc_genetic_map_server}
}
