% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_hotspot_fct_functions.R
\name{transformdata_hotspot_bc}
\alias{transformdata_hotspot_bc}
\title{Transform data for Hotspot plot for breed comparison}
\usage{
transformdata_hotspot_bc(data.trans, value, filter, infos, approach)
}
\arguments{
\item{data.trans}{list each list element represents a selected breed and contains data frame (containing: SNP,BP,D_cM,D_Theta,Hm_cM,Hm_Theta,Hf_cM,Hf_Theta,Dis)}

\item{value}{numeric contains the user specified threshold to determine the hotspot}

\item{filter}{character contains the selected chromosome}

\item{infos}{data.frame data frame containing the predefined settings and names for the selected breeds (\link{table_breed_information})}

\item{approach}{data.frame data frame containing the predefined settings and names for the selected approaches (\link{table_approach_information})}
}
\value{
The function returns a list containing: \cr
* 1. data frame containing information for hotspot plot (\link{data_plot_hotspot}-specific selected chromosome or \link{data_plot_hotspot_all} - all chromosomes is selected) \cr
* 2. data.frame for table use created for table (\link{data_table_hotspot}) \cr
* 3. list containing revealed hotspot SNP markers for each approach for the Venn diagram
}
\description{
The function used the threshold to determine hotspot SNP markers and generate corresponding output data for plot (1), table (2) and Venn diagram (3) for an selected approach and selected breeds.
1) Transform BP into Mbp and add accordingly colors, sizes and shapes. If all chromosomes are selected then an additional
column is generated "ChrTheta" - chromosome number + theta . The later is scaled in depending on the size of Theta.
2) To create the corresponding table only the detected hotspot SNP markers for each approach are used.
3) For the number of selected approaches, create corresponding sets containing the hotspot SNP markers required to create the Venn diagram.
}
\details{
function
}
\note{
The function only works for a maximum of three approaches.
}
\seealso{
* \link{mod_hotspot_server} \cr
* \link{mod_bc_hotspot_server}
}
