% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityBootstrap.R
\name{c_BootstrapStructure}
\alias{c_BootstrapStructure}
\title{Compute Residuals or Persistences on bootstrapped data by scaling them to the data from a learned clarity model}
\usage{
c_BootstrapStructure(clearned, Ybs, transform = c_Procrust)
}
\arguments{
\item{clearned}{either a ClarityScan or Clarity object. If a Clarity object, then residuals will be bootstrapped. If a ClarityScan object then Persistence will be bootstrapped, unless a specific k is given in which case Residuals are again calculated.}

\item{Ybs}{The bootstrapped matrix for which the new model will be learned}

\item{transform}{(default=c_Procrust) A function to perform a transform of the data to match the target, taking two arguments, (Ybs,target). Default: c_Procrust.}
}
\value{
A list of residuals if clearned is of class Clarity, or a list of persistences if clearned is of class ClarityScan.
}
\description{
Given a resampled matrix Ybs, a replicate of Y which was originally used to make a Clarity or ClarityScan object clearned, we:
1. Transform Ybs into YbsT, which "looks as much like Y as possible"
2. Predict YbsT using clearned
3. Compute the residual/persistence structure of this prediction

The fit is done with a k that matches that in clearned. If clearned is of class "ClarityScan" then we fit with \code{\link{Clarity_Scan}}, otherwise we fit with \code{\link{Clarity_fixedK}}. We predict the target with \code{\link{Clarity_Predict}}.
}
\seealso{
\code{\link{c_Bootstrap}} which is the recommended way to access this code.
}
