% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityBootstrap.R
\name{c_dist}
\alias{c_dist}
\title{Standard Distance function for Clarity}
\usage{
c_dist(x)
}
\arguments{
\item{x}{An N by K matrix of N subjects observed at K features}
}
\value{
An N by N matrix of the distances
}
\description{
Computes the euclidean distance between all points using as.matrix(dist(x)). Provided to provide an example for user-defined dissimilarity functions.
}
\seealso{
\code{\link{Clarity_Bootstrap}}, which takes an argument distfn which can be given this or another function that generates a similarity from feature data.
}
