% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarityCompare.R
\name{Clarity_Extract}
\alias{Clarity_Extract}
\title{Extract objects from a Clarity or ClarityScan object}
\usage{
Clarity_Extract(cscan, what = "Yresid", summary = abs, diag = 0,
  k = NULL)
}
\arguments{
\item{cscan}{A Clarity or ClarityScan object}

\item{what}{(default="Yresid") the object to extract. If a number and cscan is a ClarityScan object, returns the associated Clarity object. If it is the string "Rsq" return the square residuals.}

\item{summary}{(default=abs) any summary to apply to the result. abs is useful to create residuals that rank naturally.}

\item{diag}{(default=0) What to do with the diagonal. If NULL, nothing is done; otherwise the diagonal is set to this value. The diagonal often has inflated residuals and hence should normally be set to 0 for comparison.}

\item{k}{(default=NULL) if specified and cscan is a ClarityScan object, only the value for the specified k is returned, allowing residuals for a single k to be extracted.}
}
\value{
A list containing what for each k in the ClarityScan
}
\description{
Extract objects such as Clarity objects or residuals from a ClarityScan object. This is useful for constructing Persistences or similar. Also provides a standardized interface to extract the absolute value of residuals.
}
\examples{
\donttest{
scan=Clarity_Scan(dataraw)
e1=Clarity_Extract(scan) # Get a list containing the absolute residual matrices for each k
e2=Clarity_Extract(scan,what=5) # get the 5th Clarity object
e3=Clarity_Extract(scan,summary=I,k=5) # get the 5th Residual matrix
e4=Clarity_Extract(scan,summary=I,k=5) # get the 5th Residual matrix
e5=Clarity_Extract(scan,"Rsq",k=5) # get the 5th square residual matrix
e6=Clarity_Extract(scan,"Rsq") # Get a list containing the R squared matrix for each k
}
}
\seealso{
\code{\link{Clarity_Persistence}} to which you can pass
}
