% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_general_venn_diagram.R
\name{creating_venn}
\alias{creating_venn}
\title{Creating Venn diagram}
\usage{
creating_venn(venn_data, breed.bc, bc.venn)
}
\arguments{
\item{venn_data}{processed Venn data object (containing: x, y, name, count, geometry)}

\item{breed.bc}{vector containing the full names of selected breeds}

\item{bc.venn}{vector containing only the first letter of the selected breeds}
}
\value{
The function returns the plotted Venn diagram.
}
\description{
The function creates Venn diagram plot.
}
\details{
function
}
\seealso{
\link{mod_bc_general_server}, \link{mod_bc_genetic_map_server} and \link{mod_bc_hotspot_server}
}
