% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datamix}
\alias{datamix}
\title{Simulated Similarity dataset with modified topology for use in Clarity}
\format{A non-negative similarity matrix of dimension 100 by 100}
\source{
Uses Package \code{\link[ClaritySim]{ClaritySim}}
}
\usage{
datamix
}
\description{
A dataset containing the similarities generated by changing branch lengths and topoligical structure of \code{\link{dataraw}}

\code{library(ClaritySim)}
\code{set.seed(1)}
# Simlulate 100 objects in 10 hierarchically related clusters with 500 features
\code{mysim=simulateCoalescent(N=100,K=10,L=500,
                            alpha=0,
                            sigma0=0.0001,Amodel="uniform",
                            minedge=0.1)}
# transform the branch lengths and regenerate features and similarity
\code{mixsim<-mixCoalescent(mysim)}
\code{datamix=mixsim$Y}
}
\seealso{
Mixture matrix \code{\link{datarawA}}, Feature matrix \code{\link{datarawD}}, Similarity matrix \code{\link{dataraw}}, Similarity matrix with same topology \code{\link{datarep}}, Similarity matrix with different topology \code{\link{datamix}}.
}
\keyword{datasets}
