% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_traffic_light.R
\name{make_traffic_light}
\alias{make_traffic_light}
\title{Creating the likelihood quality signal plot}
\usage{
make_traffic_light(selected.breed, dat.tl)
}
\arguments{
\item{selected.breed}{vector containing the full names of selected breeds}

\item{dat.tl}{data frame containing the breed size information for selected breeds (\link{data_traffic_light})}
}
\value{
The function returns the plotted likelihood quality signal (ggplot object or patchwork object (containing multiple ggplot objects)).
}
\description{
The function creates the likelihood quality signal plot.
}
\details{
traffic_light
}
\seealso{
* \link{mod_general_server} \cr
* \link{mod_genetic_map_server} \cr
* \link{mod_genetic_function_server} \cr
* \link{mod_bc_general_server} \cr
* \link{mod_bc_genetic_map_server}\cr
* \link{mod_bc_genetic_function_server}
}
