% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_genetic_function_fct_functions.R
\name{makePlot_genetic_function_bc}
\alias{makePlot_genetic_function_bc}
\title{Creating plots of genetic-map functions for breed comparison}
\usage{
makePlot_genetic_function_bc(chromo, df.list, names.bc.plot, name.file)
}
\arguments{
\item{chromo}{numeric contains the selected chromosome}

\item{df.list}{list whereby the first list element contains a matrix of recombination rate and distance for marker pairs, second list element includes a matrix of X-values for all genetic-map functions,
third list element includes a matrix of Y-values for all genetic-map functions.}

\item{names.bc.plot}{character vector containing the selected breeds}

\item{name.file}{character name of output file}
}
\value{
The function returns a plotly object.
}
\description{
For a selected chromosome the curves of all genetic-map functions ("Haldane scaled","Rao","Felsenstein","Liberman & Karlin")
are generated for all selected breeds.
}
\details{
function
}
\seealso{
\link{mod_bc_genetic_function_server}
}
