% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_genetic_map_fct_functions.R
\name{makePlot_geneticMaps}
\alias{makePlot_geneticMaps}
\title{Creating the genetic map plot for selected chromosome}
\usage{
makePlot_geneticMaps(dat, name.file)
}
\arguments{
\item{dat}{data frame containing the genetic map data \link{data_plot_geneticMap}}

\item{name.file}{character contains the name for image file}
}
\value{
The function returns a plotly object.
}
\description{
The function creates the plot for a single chromosome for selected approaches (breed analysis) or the selected breeds (breed comparison)
}
\details{
function
}
\seealso{
* \link{mod_genetic_map_server} \cr
* \link{mod_bc_genetic_map_server}
}
