% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_about_the_project.R
\name{mod_about_the_project_ui}
\alias{mod_about_the_project_ui}
\alias{mod_about_the_project_server}
\title{mod_about_the_project_ui and mod_about_the_project_server}
\usage{
mod_about_the_project_ui(id)

mod_about_the_project_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module to provide all the relevant information about the project.
}
\keyword{internal}
