% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bc_genetic_function.R
\name{mod_bc_genetic_function_ui}
\alias{mod_bc_genetic_function_ui}
\alias{mod_bc_genetic_function_server}
\title{mod_bc_genetic_function_ui and mod_bc_genetic_function_server}
\usage{
mod_bc_genetic_function_ui(id)

mod_bc_genetic_function_server(
  id,
  filter,
  breed.infos,
  names.files,
  make.traff.light.bc,
  approach,
  figure_rend2_gf
)
}
\arguments{
\item{id}{module id}

\item{filter}{character contains the selected chromosome}

\item{breed.infos}{data frame containing the predefined settings and names for the selected breeds (\link{table_breed_information})}

\item{names.files}{string containing the concatenate breed names}

\item{make.traff.light.bc}{object containing the ggplot for the likelihood quality signal (\link{make_traffic_light})}

\item{approach}{data frame containing the predefined settings and names for the selected approach (\link{table_approach_information})}

\item{figure_rend2_gf}{reactive value indicates if the plot in the module has been fully rendered. Used on the main server to enable or disable "Breed comparison" sidebar elements.}
}
\description{
A shiny module to generate the outcome for the tabpanel "Genetic-map functions" for the sidebar "Breed comparison".
}
\details{
The module uses the function \link{makePlot_genetic_function_bc} for plotting.
 Moreover, the quality of the likelihood approach is shown here by means of a traffic light (also termed as
 likelihood quality signal; \link{make_traffic_light}). \link{make_barplot_bc}
}
\seealso{
* \link{makePlot_genetic_function_bc}
* \link{make_traffic_light}
* \link{make_barplot_bc}
}
