% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bc_genetic_function.R
\name{mod_bc_genetic_function_ui}
\alias{mod_bc_genetic_function_ui}
\alias{mod_bc_genetic_function_server}
\title{mod_bc_genetic_function_ui and mod_bc_genetic_function_server}
\usage{
mod_bc_genetic_function_ui(id)

mod_bc_genetic_function_server(
  id,
  filter,
  breed.select.bc,
  color.shape.def,
  names.files
)
}
\arguments{
\item{id}{module id}

\item{filter}{selected chromosome}

\item{breed.select.bc}{vector containing the names of selected breeds}

\item{color.shape.def}{data frame containing the definition for coloring, shapes, size for plots}

\item{names.files}{string containing the concatenate breed names}
}
\description{
A shiny module to generate the outcome for the tabpanel "Genetic-map functions" for the sidebar "Breed comparison".
}
\details{
The module uses the function \link{makePlot_genetic_function_bc} for plotting.
}
