% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bc_genetic_map.R
\name{mod_bc_genetic_map_ui}
\alias{mod_bc_genetic_map_ui}
\alias{mod_bc_genetic_map_server}
\title{mod_bc_genetic_map_ui and mod_bc_genetic_map_server}
\usage{
mod_bc_genetic_map_ui(id)

mod_bc_genetic_map_server(
  id,
  filter,
  geneticMap.bc,
  names.files,
  make.traff.light.bc,
  breed.infos,
  approach,
  figure_rend2_gm,
  state
)
}
\arguments{
\item{id}{module id}

\item{filter}{character contains the selected chromosome}

\item{geneticMap.bc}{list of genetic maps from the selected breeds (each list element contains a breed)}

\item{names.files}{character containing the concatenated breed names and is used as an addition to the file name}

\item{make.traff.light.bc}{object containing the ggplot for the likelihood quality signal (\link{make_traffic_light})}

\item{breed.infos}{data frame containing the predefined settings and names for the selected breeds (\link{table_breed_information})}

\item{approach}{data frame containing the predefined settings and names for the selected approach (\link{table_approach_information})}

\item{figure_rend2_gm}{reactive value indicates if the plot in the module has been fully rendered. Used on the main server to enable or disable "Breed comparison" sidebar elements.}

\item{state}{numeric indicating if the threshold slider UI should be created (0; new module ID) or not (1; reused module ID and settings retained). State of the module is tracked on the main server side.}
}
\description{
A shiny module to generate the outcome for the tabpanel "Genetic map" for the sidebar "Breed comparison".
}
\details{
Independent from the user choice regarding chromosome selection, the data are prepared for the plot
using the function \link{transformdata_genetic_map_bc}.Depending on the user's selection, when a specific chromosome is chosen, the function \link{makePlot_geneticMaps} is used to create the plot,
otherwise the function \link{makePlot_all_geneticMaps} is used to create the plot. In addition, when all chromosome is selected than the corresponding plot
is cached to increase the app performance.
This module also shows a Venn diagram of corresponding markers. The input data are prepared
using the function \link{process_venn_data}, corresponding colors for Venn sets are created using the function \link{create.colors} and plotted using the function \link{creating_venn}.
The user may select a Venn diagram subset and the
table will be reduced accordingly using the function \link{prepare_table_venn}.
Finally, the quality of the likelihood approach is also shown here by means of a traffic light (also termed as likelihood quality signal (\link{make_traffic_light})).
}
\seealso{
*\link{transformdata_genetic_map_bc} \cr
* \link{makePlot_all_geneticMaps} \cr
* \link{makePlot_geneticMaps}
* \link{process_venn_data} \cr
* \link{create.colors}
* \link{creating_venn}\cr
* \link{prepare_table_venn} \cr
* \link{make_traffic_light}
}
