% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_datasets.R
\name{mod_datasets_ui}
\alias{mod_datasets_ui}
\alias{mod_datasets_server}
\title{mod_datasets_ui and mod_datasets_server}
\usage{
mod_datasets_ui(id)

mod_datasets_server(id)
}
\arguments{
\item{id}{module id}
}
\description{
A shiny module containing the relevant information about the data sets and their external preprocessing.
}
\keyword{data}
\keyword{internal}
\keyword{internal,}
\keyword{set}
