% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_general.R
\name{mod_general_ui}
\alias{mod_general_ui}
\alias{mod_general_server}
\title{mod_general_ui and mod_general_server}
\usage{
mod_general_ui(id)

mod_general_server(input, output, session, filter)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module to generate the outcome for the tabpanel General information.
}
\details{
Both use mod_general_fct_functions were all functions are included for the graphical outputs.
}
\note{
The module uses the mod_general_fct_functions.R and the fct_general_hovering.R.
}
\keyword{internal}
