% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_general.R
\name{mod_general_ui}
\alias{mod_general_ui}
\alias{mod_general_server}
\title{mod_general_ui and mod_general_server}
\usage{
mod_general_ui(id)

mod_general_server(id, filter, breed.select, color.breed.dl)
}
\arguments{
\item{id}{module id}

\item{filter}{selected chromosome}

\item{breed.select}{name of the selected breed}

\item{color.breed.dl}{vector containing the predefined colors}
}
\description{
A shiny module to generate the outcome for the tabpanel "General information" for the sidebar "Breed analysis".
}
\details{
Within the module the data are prepared for plotting using the function \link{transformdata_general}. Moreover,
the plots are created using the function \link{scatterPlot_general_all} or \link{scatterPlot_general_selected} depending on the user selected chromosome.
Corresponding hovering information and style are generated using \link{hovering}.
}
\seealso{
\link{transformdata_general},\link{scatterPlot_general_all}, \link{scatterPlot_general_selected} and \link{hovering}
}
\keyword{general}
\keyword{internal,}
\keyword{module}
