% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_genetic_function.R
\name{mod_genetic_function_ui}
\alias{mod_genetic_function_ui}
\alias{mod_genetic_function_server}
\title{mod_genetic_function_ui and mod_genetic_function_server}
\usage{
mod_genetic_function_ui(id)

mod_genetic_function_server(input, output, session, filter)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module to generate the outcome for the tabpanel Genetic map functions.
}
\keyword{internal}
