% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_genetic_map.R
\name{mod_genetic_map_ui}
\alias{mod_genetic_map_ui}
\alias{mod_genetic_map_server}
\title{mod_genetic_map_ui and mod_genetic_map_server}
\usage{
mod_genetic_map_ui(id)

mod_genetic_map_server(
  id,
  filter,
  breed.select,
  geneticMap,
  make.traff.light,
  approach.info,
  add.approach.names
)
}
\arguments{
\item{id}{module id}

\item{filter}{character contains the selected chromosome}

\item{breed.select}{character contains the name of the selected breed}

\item{geneticMap}{data frame containing the genetic map}

\item{make.traff.light}{object containing the ggplot for the likelihood quality signal (\link{make_traffic_light})}

\item{approach.info}{data frame containing the predefined settings for selected approaches (\link{table_approach_information})}

\item{add.approach.names}{character containing the concatenated approach abbreviations and is used as an addition to the file name}
}
\description{
A shiny module to generate the outcome for the tabpanel "Genetic map" for the sidebar "Breed analysis".
}
\details{
Independent from the user choice regarding chromosome selection, the data are prepared for the plot
using the function \link{transformdata_genetic_map}. Depending on the user's selection, when a specific chromosome is chosen, the function \link{makePlot_geneticMaps} is used to create the plot,
otherwise the function \link{makePlot_all_geneticMaps} is used to create the plot. In addition, when all chromosome is selected than the corresponding plot
is cached to increase the app performance.
Finally, the quality of the likelihood approach is also shown here by means of a traffic light
 (also termed as likelihood quality signal (\link{make_traffic_light})).
}
\seealso{
* \link{transformdata_genetic_map} \cr
* \link{makePlot_geneticMaps} \cr
* \link{makePlot_all_geneticMaps} \cr
* \link{make_traffic_light}
}
\keyword{genetic}
\keyword{internal,}
\keyword{map}
