% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_genetic_map.R
\name{mod_genetic_map_ui}
\alias{mod_genetic_map_ui}
\alias{mod_genetic_map_server}
\title{mod_genetic_map_ui and mod_genetic_map_server}
\usage{
mod_genetic_map_ui(id)

mod_genetic_map_server(input, output, session, filter)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{filter}{external depends on choice by the chromosome}
}
\description{
A shiny Module to generate the outcome for the tabpanel Genetic map.
}
\note{
The module uses the mod_genetic_map_fct_functions.R.
}
\keyword{internal}
