% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_hotspot.R
\name{mod_hotspot_ui}
\alias{mod_hotspot_ui}
\alias{mod_hotspot_server}
\title{mod_hotspot_ui and mod_hotspot_server}
\usage{
mod_hotspot_ui(id)

mod_hotspot_server(id, filter, breed.select, color.shape.def)
}
\arguments{
\item{id}{module id}

\item{filter}{character contains the selected chromosome}

\item{breed.select}{character contains the name of the selected breed}

\item{color.shape.def}{data frame containing the definition for coloring, shapes, size for plots}
}
\description{
A shiny module to generate the outcome for the tabpanel "General information" for the sidebar "Breed analysis".
}
\details{
Within the module the corresponding data are prepared using the function \link{transformdata_hotspot}. The prepared data are used to plot the hotspot figure using the function
\link{scatterPlot_hotspot} or \link{scatterPlot_hotspot_all} depending on user chromosome choice. The hover information are obtained from the function \link{hovering}.
}
\seealso{
\link{transformdata_hotspot}, \link{scatterPlot_hotspot}, \link{scatterPlot_hotspot_all} and \link{hovering}
}
\keyword{hotspot}
\keyword{internal}
\keyword{internal,}
