% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_methodology.R
\name{mod_methodology_ui}
\alias{mod_methodology_ui}
\alias{mod_methodology_server}
\title{mod_methodology_ui and mod_methodology_server}
\usage{
mod_methodology_ui(id)

mod_methodology_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module containing all relevant information about XXX.
}
\keyword{internal}
