% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_misplaced.R
\name{mod_misplaced_ui}
\alias{mod_misplaced_ui}
\alias{mod_misplaced_server}
\title{mod_misplaced_ui and mod_misplaced_server}
\usage{
mod_misplaced_ui(id)

mod_misplaced_server(id)
}
\arguments{
\item{id}{module id}
}
\description{
A shiny module to generate the outcome for the misplaced markers.
}
\keyword{internal,}
\keyword{marker}
\keyword{misplaced}
