% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_general_venn_diagram.R
\name{prepare_table_venn}
\alias{prepare_table_venn}
\title{Creating a reduced table}
\usage{
prepare_table_venn(venn.dat, venn, click, table.bc)
}
\arguments{
\item{venn.dat}{processed Venn data object (containing: x, y, name, count, geometry)}

\item{venn}{object of class "Venn"}

\item{click}{selected Venn diagram subset}

\item{table.bc}{data frame}
}
\value{
The function returns a list: 1. element is a table containing the unique set and 2. element contains which Venn subset was selected.
}
\description{
The function creates a reduced table based on the selected Venn diagram subset. The function is used by \link{mod_bc_genetic_map_server}.
}
\details{
function
}
\seealso{
\link{mod_bc_genetic_map_server} and  \link{mod_bc_hotspot_server}
}
