% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_general_venn_diagram.R
\name{prepare_table_venn}
\alias{prepare_table_venn}
\title{Creating a reduced table}
\usage{
prepare_table_venn(venn.dat, venn, click, table.bc)
}
\arguments{
\item{venn.dat}{object containing the processed Venn data (containing the plot data; created in \link{process_venn_data})}

\item{venn}{object of class "Venn"}

\item{click}{list containing the Venn diagram information from the user mouse click}

\item{table.bc}{data frame containing the information which are used for the output table depending from the called module \cr
(\link{data_table_hotspot} for hotspot detection, or \link{data_table_geneticMap_bc} for genetic map (only breed comparison))}
}
\value{
The function returns a list:\cr
1. element contains the reduced table accordingly to the selected unique Venn set \cr
2. element contains which Venn subset was selected.
}
\description{
The function creates a reduced table based on the selected Venn diagram SNP marker subset.
}
\details{
function
}
\seealso{
* \link{mod_bc_genetic_map_server}  \cr
* \link{mod_bc_hotspot_server} \cr
* \link{mod_hotspot_server}
}
