% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_hotspot_fct_functions.R
\name{scatterPlot_hotspot}
\alias{scatterPlot_hotspot}
\title{Creating the hotspot scatter plot for a specific chromosomome}
\usage{
scatterPlot_hotspot(dat1, file.name)
}
\arguments{
\item{dat1}{data frame \link{data_plot_hotspot}}

\item{file.name}{character containing the file name}
}
\value{
The function returns a plotly object.
}
\description{
The function creates the hotspot plot for a specific selected chromosome.
}
\details{
function
}
\seealso{
* \link{mod_hotspot_server}
* \link{mod_bc_hotspot_server}
}
