% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_hotspot_fct_functions.R
\name{scatterPlot_hotspot_all}
\alias{scatterPlot_hotspot_all}
\title{Creating the Hotspot scatter plot for all chromosomomes}
\usage{
scatterPlot_hotspot_all(dat1, ranges, scale, legend1)
}
\arguments{
\item{dat1}{data frame (containing: Chr, SNP, cM, BP, Theta, Dis, ord, size, coloring, and shape)}

\item{ranges}{contains brush information}

\item{scale}{number to scale the plot}

\item{legend1}{data frame containing the legend information (contains: label, breaks, shape, and ord)}
}
\value{
The function returns a ggplot object.
}
\description{
The function creates the hotspot plot for all chromosomes.
}
\details{
function

Note: legend1, the column breaks contains the color for the corresponding label, and column ord contains numbers used for sorting to keep the sequence (corresponds to ord in dat1).
}
\seealso{
\link{mod_hotspot_server} and \link{mod_bc_hotspot_server}
}
