% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_general_fct_functions.R
\name{transformdata_general}
\alias{transformdata_general}
\title{Transform data for breed analysis}
\usage{
transformdata_general(data1)
}
\arguments{
\item{data1}{is a data frame}
}
\value{
The function returns a data frame with three columns: basepairs (bp), centiMorgan (cM) and chromosome (chr).
}
\description{
The function transforms the data for plotting: basepairs (bp) into megabasepairs (Mbp) and Morgan (M) into centiMorgan (cM).
}
\details{
function
}
\seealso{
\link{mod_general_server}
}
