% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_general_fct_functions.R
\name{transformdata_general_bc}
\alias{transformdata_general_bc}
\title{Transform data for breed comparison}
\usage{
transformdata_general_bc(data1, what, breed.select.bc, colo)
}
\arguments{
\item{data1}{data frame}

\item{what}{name of the approach
\itemize{
\item det - is used for deterministic approach
\item lik - is used for likelihood-based approach
}}

\item{breed.select.bc}{vector containing the selected breeds}

\item{colo}{vector containing the predefined colors for each breed}
}
\value{
The function returns a data frame with four columns: basepair (bp), centiMorgan (cM), chromosome (chr), breed and color.
}
\description{
The function transforms the data for plotting: basepairs (bp) into megabasepairs (Mbp) and Morgan (M) into centiMorgan (cM) for selected breeds and adds
breed specific colors.
}
\details{
function
}
\seealso{
\link{mod_bc_general_server}
}
