% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_general_fct_functions.R
\name{transformdata_general_bc}
\alias{transformdata_general_bc}
\title{Transform data for breed comparison}
\usage{
transformdata_general_bc(data1, breed.infos, filter)
}
\arguments{
\item{data1}{list each list element represents a chosen breed and contains a data frame with genetic map summary for selected approach.}

\item{breed.infos}{data frame containing the predefined settings and names for the selected breeds (\link{table_breed_information})}

\item{filter}{character contains the selected chromosome}
}
\value{
The function returns a data frame (\link{data_plot_general})
}
\description{
The function transforms the data for plotting:
1) basepairs (bp) into megabasepairs (Mbp) and Morgan (M) into centiMorgan (cM) for selected breeds,
2) adds breed specific colors and breed name, and
3) adds size values ("10" not selected chromosome and "18" selected chromosome; 14 when all chromosomes are selected)
}
\details{
function
}
\seealso{
\link{mod_bc_general_server}
}
