% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_genetic_map_fct_functions.R
\name{transformdata_genetic_map}
\alias{transformdata_genetic_map}
\title{Prepare genetic map data for plotting for breed analysis}
\usage{
transformdata_genetic_map(input, approach.info)
}
\arguments{
\item{input}{data frame containing the genetic map information}

\item{approach.info}{data frame containing the predefined settings for selected approaches (\link{table_approach_information})}
}
\value{
The function returns list containing the desired information in data frame (\link{data_plot_geneticMap}) for each chromosome.
}
\description{
The function prepares the genetic map information for a single breed and selected approaches.
}
\details{
function
}
\seealso{
\link{mod_genetic_map_server}
}
