% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_genetic_map_fct_functions.R
\name{transformdata_genetic_map_bc}
\alias{transformdata_genetic_map_bc}
\title{Prepare genetic map data for plotting for breed comparison}
\usage{
transformdata_genetic_map_bc(input, breed.infos, approach)
}
\arguments{
\item{input}{list of genetic maps for the selected breeds}

\item{breed.infos}{data frame containing the predefined settings for selected breeds (\link{table_breed_information})}

\item{approach}{character containing the selected approach}
}
\value{
The function returns a list containing the desired information in data frame for each chromosome (\link{data_plot_geneticMap})
}
\description{
The function prepares the genetic map information for a single approach and chosen breeds.
}
\details{
function
}
\seealso{
\link{mod_bc_genetic_map_server}
}
