% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_hotspot_fct_functions.R
\name{transformdata_hotspot}
\alias{transformdata_hotspot}
\title{Transform data for Hotspot plot}
\usage{
transformdata_hotspot(
  data.trans,
  value = value,
  color1 = color1,
  shape1 = shape1,
  ord = ord
)
}
\arguments{
\item{data.trans}{data frame (containing: Chr, SNP, cM, BP, Theta, Dis)}

\item{value}{contains the user specified threshold to determine the hotspot}

\item{color1}{vector containing the predefined used colors}

\item{shape1}{vector containing the predefined used shapes}

\item{ord}{vector containing the predefined numbering}
}
\value{
The function returns a data frame (containing: Chr, SNP, cM, BP, Theta, Dis, ord, size, coloring, and shape).
}
\description{
The function used the threshold to determine hotspot markers and add accordingly the colors, shapes and numbering.
}
\details{
function
}
\seealso{
\link{mod_hotspot_server} and \link{mod_bc_hotspot_server}
}
